Get a list of country names

This function returns a vector of country names in different nomenclatures.

Usage

list_countries(nomenclature = "name_en")

Arguments

nomenclature

String indicating the nomenclature from which the list of countries should be taken. Not all countries are present in all nomenclatures, for example Taiwan is not recognised by the UN, so it will not be returned with "WTO_en". The function accepts any of the nomenclatures supported country_name. For a list of accepted values, refer to this page. The default is name_en, which is the English list of names in the ISO standard 3166.

Value

A vector of country names in the desired nomenclature.

See also

Examples

list_countries("ISO3")
#>   [1] "AFG" "ALA" "ALB" "DZA" "ASM" "AND" "AGO" "AIA" "ATA" "ATG" "ARG" "ARM"
#>  [13] "ABW" "AUS" "AUT" "AZE" "BHS" "BHR" "BGD" "BRB" "BLR" "BEL" "BLZ" "BEN"
#>  [25] "BMU" "BTN" "BOL" "BES" "BIH" "BWA" "BVT" "BRA" "IOT" "VGB" "BRN" "BGR"
#>  [37] "BFA" "BDI" "CPV" "KHM" "CMR" "CAN" "CYM" "CAF" "TCD" "CHL" "CHN" "HKG"
#>  [49] "MAC" "CXR" "CCK" "COL" "COM" "COG" "COK" "CRI" "CIV" "HRV" "CUB" "CUW"
#>  [61] "CYP" "CZE" "PRK" "COD" "DNK" "DJI" "DMA" "DOM" "ECU" "EGY" "SLV" "GNQ"
#>  [73] "ERI" "EST" "SWZ" "ETH" "FLK" "FRO" "FJI" "FIN" "FRA" "GUF" "PYF" "ATF"
#>  [85] "GAB" "GMB" "GEO" "DEU" "GHA" "GIB" "GRC" "GRL" "GRD" "GLP" "GUM" "GTM"
#>  [97] "GGY" "GIN" "GNB" "GUY" "HTI" "HMD" "VAT" "HND" "HUN" "ISL" "IND" "IDN"
#> [109] "IRN" "IRQ" "IRL" "IMN" "ISR" "ITA" "JAM" "JPN" "JEY" "JOR" "KAZ" "KEN"
#> [121] "KIR" "KWT" "KGZ" "LAO" "LVA" "LBN" "LSO" "LBR" "LBY" "LIE" "LTU" "LUX"
#> [133] "MDG" "MWI" "MYS" "MDV" "MLI" "MLT" "MHL" "MTQ" "MRT" "MUS" "MYT" "MEX"
#> [145] "FSM" "MCO" "MNG" "MNE" "MSR" "MAR" "MOZ" "MMR" "NAM" "NRU" "NPL" "NLD"
#> [157] "NCL" "NZL" "NIC" "NER" "NGA" "NIU" "NFK" "MKD" "MNP" "NOR" "OMN" "PAK"
#> [169] "PLW" "PAN" "PNG" "PRY" "PER" "PHL" "PCN" "POL" "PRT" "PRI" "QAT" "KOR"
#> [181] "MDA" "REU" "ROU" "RUS" "RWA" "BLM" "SHN" "KNA" "LCA" "MAF" "SPM" "VCT"
#> [193] "WSM" "SMR" "STP" "SAU" "SEN" "SRB" "SYC" "SLE" "SGP" "SXM" "SVK" "SVN"
#> [205] "SLB" "SOM" "ZAF" "SGS" "SSD" "ESP" "LKA" "PSE" "SDN" "SUR" "SJM" "SWE"
#> [217] "CHE" "SYR" "TJK" "THA" "TLS" "TGO" "TKL" "TON" "TTO" "TUN" "TUR" "TKM"
#> [229] "TCA" "TUV" "UGA" "UKR" "ARE" "GBR" "TZA" "UMI" "USA" "VIR" "URY" "UZB"
#> [241] "VUT" "VEN" "VNM" "WLF" "ESH" "YEM" "ZMB" "ZWE" "TWN"
list_countries("UN_en")
#>   [1] "Afghanistan"                                         
#>   [2] "Åland Islands"                                       
#>   [3] "Albania"                                             
#>   [4] "Algeria"                                             
#>   [5] "American Samoa"                                      
#>   [6] "Andorra"                                             
#>   [7] "Angola"                                              
#>   [8] "Anguilla"                                            
#>   [9] "Antarctica"                                          
#>  [10] "Antigua and Barbuda"                                 
#>  [11] "Argentina"                                           
#>  [12] "Armenia"                                             
#>  [13] "Aruba"                                               
#>  [14] "Australia"                                           
#>  [15] "Austria"                                             
#>  [16] "Azerbaijan"                                          
#>  [17] "Bahamas"                                             
#>  [18] "Bahrain"                                             
#>  [19] "Bangladesh"                                          
#>  [20] "Barbados"                                            
#>  [21] "Belarus"                                             
#>  [22] "Belgium"                                             
#>  [23] "Belize"                                              
#>  [24] "Benin"                                               
#>  [25] "Bermuda"                                             
#>  [26] "Bhutan"                                              
#>  [27] "Bolivia (Plurinational State of)"                    
#>  [28] "Bonaire, Sint Eustatius and Saba"                    
#>  [29] "Bosnia and Herzegovina"                              
#>  [30] "Botswana"                                            
#>  [31] "Bouvet Island"                                       
#>  [32] "Brazil"                                              
#>  [33] "British Indian Ocean Territory"                      
#>  [34] "British Virgin Islands"                              
#>  [35] "Brunei Darussalam"                                   
#>  [36] "Bulgaria"                                            
#>  [37] "Burkina Faso"                                        
#>  [38] "Burundi"                                             
#>  [39] "Cabo Verde"                                          
#>  [40] "Cambodia"                                            
#>  [41] "Cameroon"                                            
#>  [42] "Canada"                                              
#>  [43] "Cayman Islands"                                      
#>  [44] "Central African Republic"                            
#>  [45] "Chad"                                                
#>  [46] "Chile"                                               
#>  [47] "China"                                               
#>  [48] "China, Hong Kong Special Administrative Region"      
#>  [49] "China, Macao Special Administrative Region"          
#>  [50] "Christmas Island"                                    
#>  [51] "Cocos (Keeling) Islands"                             
#>  [52] "Colombia"                                            
#>  [53] "Comoros"                                             
#>  [54] "Congo"                                               
#>  [55] "Cook Islands"                                        
#>  [56] "Costa Rica"                                          
#>  [57] "Côte d’Ivoire"                                       
#>  [58] "Croatia"                                             
#>  [59] "Cuba"                                                
#>  [60] "Curaçao"                                             
#>  [61] "Cyprus"                                              
#>  [62] "Czechia"                                             
#>  [63] "Democratic People's Republic of Korea"               
#>  [64] "Democratic Republic of the Congo"                    
#>  [65] "Denmark"                                             
#>  [66] "Djibouti"                                            
#>  [67] "Dominica"                                            
#>  [68] "Dominican Republic"                                  
#>  [69] "Ecuador"                                             
#>  [70] "Egypt"                                               
#>  [71] "El Salvador"                                         
#>  [72] "Equatorial Guinea"                                   
#>  [73] "Eritrea"                                             
#>  [74] "Estonia"                                             
#>  [75] "Eswatini"                                            
#>  [76] "Ethiopia"                                            
#>  [77] "Falkland Islands (Malvinas)"                         
#>  [78] "Faroe Islands"                                       
#>  [79] "Fiji"                                                
#>  [80] "Finland"                                             
#>  [81] "France"                                              
#>  [82] "French Guiana"                                       
#>  [83] "French Polynesia"                                    
#>  [84] "French Southern Territories"                         
#>  [85] "Gabon"                                               
#>  [86] "Gambia"                                              
#>  [87] "Georgia"                                             
#>  [88] "Germany"                                             
#>  [89] "Ghana"                                               
#>  [90] "Gibraltar"                                           
#>  [91] "Greece"                                              
#>  [92] "Greenland"                                           
#>  [93] "Grenada"                                             
#>  [94] "Guadeloupe"                                          
#>  [95] "Guam"                                                
#>  [96] "Guatemala"                                           
#>  [97] "Guernsey"                                            
#>  [98] "Guinea"                                              
#>  [99] "Guinea-Bissau"                                       
#> [100] "Guyana"                                              
#> [101] "Haiti"                                               
#> [102] "Heard Island and McDonald Islands"                   
#> [103] "Holy See"                                            
#> [104] "Honduras"                                            
#> [105] "Hungary"                                             
#> [106] "Iceland"                                             
#> [107] "India"                                               
#> [108] "Indonesia"                                           
#> [109] "Iran (Islamic Republic of)"                          
#> [110] "Iraq"                                                
#> [111] "Ireland"                                             
#> [112] "Isle of Man"                                         
#> [113] "Israel"                                              
#> [114] "Italy"                                               
#> [115] "Jamaica"                                             
#> [116] "Japan"                                               
#> [117] "Jersey"                                              
#> [118] "Jordan"                                              
#> [119] "Kazakhstan"                                          
#> [120] "Kenya"                                               
#> [121] "Kiribati"                                            
#> [122] "Kuwait"                                              
#> [123] "Kyrgyzstan"                                          
#> [124] "Lao People's Democratic Republic"                    
#> [125] "Latvia"                                              
#> [126] "Lebanon"                                             
#> [127] "Lesotho"                                             
#> [128] "Liberia"                                             
#> [129] "Libya"                                               
#> [130] "Liechtenstein"                                       
#> [131] "Lithuania"                                           
#> [132] "Luxembourg"                                          
#> [133] "Madagascar"                                          
#> [134] "Malawi"                                              
#> [135] "Malaysia"                                            
#> [136] "Maldives"                                            
#> [137] "Mali"                                                
#> [138] "Malta"                                               
#> [139] "Marshall Islands"                                    
#> [140] "Martinique"                                          
#> [141] "Mauritania"                                          
#> [142] "Mauritius"                                           
#> [143] "Mayotte"                                             
#> [144] "Mexico"                                              
#> [145] "Micronesia (Federated States of)"                    
#> [146] "Monaco"                                              
#> [147] "Mongolia"                                            
#> [148] "Montenegro"                                          
#> [149] "Montserrat"                                          
#> [150] "Morocco"                                             
#> [151] "Mozambique"                                          
#> [152] "Myanmar"                                             
#> [153] "Namibia"                                             
#> [154] "Nauru"                                               
#> [155] "Nepal"                                               
#> [156] "Netherlands"                                         
#> [157] "New Caledonia"                                       
#> [158] "New Zealand"                                         
#> [159] "Nicaragua"                                           
#> [160] "Niger"                                               
#> [161] "Nigeria"                                             
#> [162] "Niue"                                                
#> [163] "Norfolk Island"                                      
#> [164] "North Macedonia"                                     
#> [165] "Northern Mariana Islands"                            
#> [166] "Norway"                                              
#> [167] "Oman"                                                
#> [168] "Pakistan"                                            
#> [169] "Palau"                                               
#> [170] "Panama"                                              
#> [171] "Papua New Guinea"                                    
#> [172] "Paraguay"                                            
#> [173] "Peru"                                                
#> [174] "Philippines"                                         
#> [175] "Pitcairn"                                            
#> [176] "Poland"                                              
#> [177] "Portugal"                                            
#> [178] "Puerto Rico"                                         
#> [179] "Qatar"                                               
#> [180] "Republic of Korea"                                   
#> [181] "Republic of Moldova"                                 
#> [182] "Réunion"                                             
#> [183] "Romania"                                             
#> [184] "Russian Federation"                                  
#> [185] "Rwanda"                                              
#> [186] "Saint Barthélemy"                                    
#> [187] "Saint Helena"                                        
#> [188] "Saint Kitts and Nevis"                               
#> [189] "Saint Lucia"                                         
#> [190] "Saint Martin (French Part)"                          
#> [191] "Saint Pierre and Miquelon"                           
#> [192] "Saint Vincent and the Grenadines"                    
#> [193] "Samoa"                                               
#> [194] "San Marino"                                          
#> [195] "Sao Tome and Principe"                               
#> [196] "Saudi Arabia"                                        
#> [197] "Senegal"                                             
#> [198] "Serbia"                                              
#> [199] "Seychelles"                                          
#> [200] "Sierra Leone"                                        
#> [201] "Singapore"                                           
#> [202] "Sint Maarten (Dutch part)"                           
#> [203] "Slovakia"                                            
#> [204] "Slovenia"                                            
#> [205] "Solomon Islands"                                     
#> [206] "Somalia"                                             
#> [207] "South Africa"                                        
#> [208] "South Georgia and the South Sandwich Islands"        
#> [209] "South Sudan"                                         
#> [210] "Spain"                                               
#> [211] "Sri Lanka"                                           
#> [212] "State of Palestine"                                  
#> [213] "Sudan"                                               
#> [214] "Suriname"                                            
#> [215] "Svalbard and Jan Mayen Islands"                      
#> [216] "Sweden"                                              
#> [217] "Switzerland"                                         
#> [218] "Syrian Arab Republic"                                
#> [219] "Tajikistan"                                          
#> [220] "Thailand"                                            
#> [221] "Timor-Leste"                                         
#> [222] "Togo"                                                
#> [223] "Tokelau"                                             
#> [224] "Tonga"                                               
#> [225] "Trinidad and Tobago"                                 
#> [226] "Tunisia"                                             
#> [227] "Türkiye"                                             
#> [228] "Turkmenistan"                                        
#> [229] "Turks and Caicos Islands"                            
#> [230] "Tuvalu"                                              
#> [231] "Uganda"                                              
#> [232] "Ukraine"                                             
#> [233] "United Arab Emirates"                                
#> [234] "United Kingdom of Great Britain and Northern Ireland"
#> [235] "United Republic of Tanzania"                         
#> [236] "United States Minor Outlying Islands"                
#> [237] "United States of America"                            
#> [238] "United States Virgin Islands"                        
#> [239] "Uruguay"                                             
#> [240] "Uzbekistan"                                          
#> [241] "Vanuatu"                                             
#> [242] "Venezuela (Bolivarian Republic of)"                  
#> [243] "Viet Nam"                                            
#> [244] "Wallis and Futuna Islands"                           
#> [245] "Western Sahara"                                      
#> [246] "Yemen"                                               
#> [247] "Zambia"                                              
#> [248] "Zimbabwe"                                            
list_countries()
#>   [1] "Afghanistan"                                         
#>   [2] "Åland Islands"                                       
#>   [3] "Albania"                                             
#>   [4] "Algeria"                                             
#>   [5] "American Samoa"                                      
#>   [6] "Andorra"                                             
#>   [7] "Angola"                                              
#>   [8] "Anguilla"                                            
#>   [9] "Antarctica"                                          
#>  [10] "Antigua and Barbuda"                                 
#>  [11] "Argentina"                                           
#>  [12] "Armenia"                                             
#>  [13] "Aruba"                                               
#>  [14] "Australia"                                           
#>  [15] "Austria"                                             
#>  [16] "Azerbaijan"                                          
#>  [17] "Bahamas"                                             
#>  [18] "Bahrain"                                             
#>  [19] "Bangladesh"                                          
#>  [20] "Barbados"                                            
#>  [21] "Belarus"                                             
#>  [22] "Belgium"                                             
#>  [23] "Belize"                                              
#>  [24] "Benin"                                               
#>  [25] "Bermuda"                                             
#>  [26] "Bhutan"                                              
#>  [27] "Bolivia (Plurinational State of)"                    
#>  [28] "Bonaire, Sint Eustatius and Saba"                    
#>  [29] "Bosnia and Herzegovina"                              
#>  [30] "Botswana"                                            
#>  [31] "Bouvet Island"                                       
#>  [32] "Brazil"                                              
#>  [33] "British Indian Ocean Territory"                      
#>  [34] "Virgin Islands (British)"                            
#>  [35] "Brunei Darussalam"                                   
#>  [36] "Bulgaria"                                            
#>  [37] "Burkina Faso"                                        
#>  [38] "Burundi"                                             
#>  [39] "Cabo Verde"                                          
#>  [40] "Cambodia"                                            
#>  [41] "Cameroon"                                            
#>  [42] "Canada"                                              
#>  [43] "Cayman Islands"                                      
#>  [44] "Central African Republic"                            
#>  [45] "Chad"                                                
#>  [46] "Chile"                                               
#>  [47] "China"                                               
#>  [48] "Hong Kong"                                           
#>  [49] "Macao"                                               
#>  [50] "Christmas Island"                                    
#>  [51] "Cocos (Keeling) Islands"                             
#>  [52] "Colombia"                                            
#>  [53] "Comoros"                                             
#>  [54] "Congo"                                               
#>  [55] "Cook Islands"                                        
#>  [56] "Costa Rica"                                          
#>  [57] "Côte d'Ivoire"                                       
#>  [58] "Croatia"                                             
#>  [59] "Cuba"                                                
#>  [60] "Curaçao"                                             
#>  [61] "Cyprus"                                              
#>  [62] "Czechia"                                             
#>  [63] "Korea (Democratic People's Republic of)"             
#>  [64] "Congo, Democratic Republic of the"                   
#>  [65] "Denmark"                                             
#>  [66] "Djibouti"                                            
#>  [67] "Dominica"                                            
#>  [68] "Dominican Republic"                                  
#>  [69] "Ecuador"                                             
#>  [70] "Egypt"                                               
#>  [71] "El Salvador"                                         
#>  [72] "Equatorial Guinea"                                   
#>  [73] "Eritrea"                                             
#>  [74] "Estonia"                                             
#>  [75] "Eswatini"                                            
#>  [76] "Ethiopia"                                            
#>  [77] "Falkland Islands (Malvinas)"                         
#>  [78] "Faroe Islands"                                       
#>  [79] "Fiji"                                                
#>  [80] "Finland"                                             
#>  [81] "France"                                              
#>  [82] "French Guiana"                                       
#>  [83] "French Polynesia"                                    
#>  [84] "French Southern Territories"                         
#>  [85] "Gabon"                                               
#>  [86] "Gambia"                                              
#>  [87] "Georgia"                                             
#>  [88] "Germany"                                             
#>  [89] "Ghana"                                               
#>  [90] "Gibraltar"                                           
#>  [91] "Greece"                                              
#>  [92] "Greenland"                                           
#>  [93] "Grenada"                                             
#>  [94] "Guadeloupe"                                          
#>  [95] "Guam"                                                
#>  [96] "Guatemala"                                           
#>  [97] "Guernsey"                                            
#>  [98] "Guinea"                                              
#>  [99] "Guinea-Bissau"                                       
#> [100] "Guyana"                                              
#> [101] "Haiti"                                               
#> [102] "Heard Island and McDonald Islands"                   
#> [103] "Holy See"                                            
#> [104] "Honduras"                                            
#> [105] "Hungary"                                             
#> [106] "Iceland"                                             
#> [107] "India"                                               
#> [108] "Indonesia"                                           
#> [109] "Iran (Islamic Republic of)"                          
#> [110] "Iraq"                                                
#> [111] "Ireland"                                             
#> [112] "Isle of Man"                                         
#> [113] "Israel"                                              
#> [114] "Italy"                                               
#> [115] "Jamaica"                                             
#> [116] "Japan"                                               
#> [117] "Jersey"                                              
#> [118] "Jordan"                                              
#> [119] "Kazakhstan"                                          
#> [120] "Kenya"                                               
#> [121] "Kiribati"                                            
#> [122] "Kuwait"                                              
#> [123] "Kyrgyzstan"                                          
#> [124] "Lao People's Democratic Republic"                    
#> [125] "Latvia"                                              
#> [126] "Lebanon"                                             
#> [127] "Lesotho"                                             
#> [128] "Liberia"                                             
#> [129] "Libya"                                               
#> [130] "Liechtenstein"                                       
#> [131] "Lithuania"                                           
#> [132] "Luxembourg"                                          
#> [133] "Madagascar"                                          
#> [134] "Malawi"                                              
#> [135] "Malaysia"                                            
#> [136] "Maldives"                                            
#> [137] "Mali"                                                
#> [138] "Malta"                                               
#> [139] "Marshall Islands"                                    
#> [140] "Martinique"                                          
#> [141] "Mauritania"                                          
#> [142] "Mauritius"                                           
#> [143] "Mayotte"                                             
#> [144] "Mexico"                                              
#> [145] "Micronesia (Federated States of)"                    
#> [146] "Monaco"                                              
#> [147] "Mongolia"                                            
#> [148] "Montenegro"                                          
#> [149] "Montserrat"                                          
#> [150] "Morocco"                                             
#> [151] "Mozambique"                                          
#> [152] "Myanmar"                                             
#> [153] "Namibia"                                             
#> [154] "Nauru"                                               
#> [155] "Nepal"                                               
#> [156] "Netherlands"                                         
#> [157] "New Caledonia"                                       
#> [158] "New Zealand"                                         
#> [159] "Nicaragua"                                           
#> [160] "Niger"                                               
#> [161] "Nigeria"                                             
#> [162] "Niue"                                                
#> [163] "Norfolk Island"                                      
#> [164] "North Macedonia"                                     
#> [165] "Northern Mariana Islands"                            
#> [166] "Norway"                                              
#> [167] "Oman"                                                
#> [168] "Pakistan"                                            
#> [169] "Palau"                                               
#> [170] "Panama"                                              
#> [171] "Papua New Guinea"                                    
#> [172] "Paraguay"                                            
#> [173] "Peru"                                                
#> [174] "Philippines"                                         
#> [175] "Pitcairn"                                            
#> [176] "Poland"                                              
#> [177] "Portugal"                                            
#> [178] "Puerto Rico"                                         
#> [179] "Qatar"                                               
#> [180] "Korea, Republic of"                                  
#> [181] "Moldova, Republic of"                                
#> [182] "Réunion"                                             
#> [183] "Romania"                                             
#> [184] "Russian Federation"                                  
#> [185] "Rwanda"                                              
#> [186] "Saint Barthélemy"                                    
#> [187] "Saint Helena, Ascension and Tristan da Cunha"        
#> [188] "Saint Kitts and Nevis"                               
#> [189] "Saint Lucia"                                         
#> [190] "Saint Martin (French part)"                          
#> [191] "Saint Pierre and Miquelon"                           
#> [192] "Saint Vincent and the Grenadines"                    
#> [193] "Samoa"                                               
#> [194] "San Marino"                                          
#> [195] "Sao Tome and Principe"                               
#> [196] "Saudi Arabia"                                        
#> [197] "Senegal"                                             
#> [198] "Serbia"                                              
#> [199] "Seychelles"                                          
#> [200] "Sierra Leone"                                        
#> [201] "Singapore"                                           
#> [202] "Sint Maarten (Dutch part)"                           
#> [203] "Slovakia"                                            
#> [204] "Slovenia"                                            
#> [205] "Solomon Islands"                                     
#> [206] "Somalia"                                             
#> [207] "South Africa"                                        
#> [208] "South Georgia and the South Sandwich Islands"        
#> [209] "South Sudan"                                         
#> [210] "Spain"                                               
#> [211] "Sri Lanka"                                           
#> [212] "Palestine, State of"                                 
#> [213] "Sudan"                                               
#> [214] "Suriname"                                            
#> [215] "Svalbard and Jan Mayen"                              
#> [216] "Sweden"                                              
#> [217] "Switzerland"                                         
#> [218] "Syrian Arab Republic"                                
#> [219] "Tajikistan"                                          
#> [220] "Thailand"                                            
#> [221] "Timor-Leste"                                         
#> [222] "Togo"                                                
#> [223] "Tokelau"                                             
#> [224] "Tonga"                                               
#> [225] "Trinidad and Tobago"                                 
#> [226] "Tunisia"                                             
#> [227] "Turkey"                                              
#> [228] "Turkmenistan"                                        
#> [229] "Turks and Caicos Islands"                            
#> [230] "Tuvalu"                                              
#> [231] "Uganda"                                              
#> [232] "Ukraine"                                             
#> [233] "United Arab Emirates"                                
#> [234] "United Kingdom of Great Britain and Northern Ireland"
#> [235] "Tanzania, United Republic of"                        
#> [236] "United States Minor Outlying Islands"                
#> [237] "United States of America"                            
#> [238] "Virgin Islands (U.S.)"                               
#> [239] "Uruguay"                                             
#> [240] "Uzbekistan"                                          
#> [241] "Vanuatu"                                             
#> [242] "Venezuela (Bolivarian Republic of)"                  
#> [243] "Viet Nam"                                            
#> [244] "Wallis and Futuna"                                   
#> [245] "Western Sahara"                                      
#> [246] "Yemen"                                               
#> [247] "Zambia"                                              
#> [248] "Zimbabwe"                                            
#> [249] "Taiwan, Province of China"