Output random country names

This function returns the mode of vectors. That is to say, for any given vector of values, it returns the value that appears most frequently. The function works with strings, numerical and mixed inputs. NA values are treated as distinct values.

Usage

random_countries(n, replace = FALSE, nomenclature = "name_en", seed = NULL)

Arguments

n

Number of desired (pseudo)random country names.

replace

Logical value indicating whether sampling should be with replacement.

nomenclature

Nomenclature from which the list of countries should be taken. Not all countries are present in all nomenclature, for example Taiwan is not recognised by the UN, so it will not be returned with "WTO_en". The function accept any of the nomenclatures of country_name. For a list of accepted values, refer to this page. The default is name_en, which is the English list of names in the ISO standard 3166.

seed

Single numerical value to be used as seed.

Value

A vector of n (pseudo)random country names.

See also

Examples

random_countries(10)
#>  [1] "Guatemala"                   "Western Sahara"             
#>  [3] "Monaco"                      "Falkland Islands (Malvinas)"
#>  [5] "Turkmenistan"                "Ghana"                      
#>  [7] "Eritrea"                     "Hong Kong"                  
#>  [9] "Egypt"                       "French Guiana"              
random_countries(n = 500, replace = TRUE)
#>   [1] "Libya"                                               
#>   [2] "Norfolk Island"                                      
#>   [3] "Greece"                                              
#>   [4] "Saint Pierre and Miquelon"                           
#>   [5] "Syrian Arab Republic"                                
#>   [6] "Costa Rica"                                          
#>   [7] "Qatar"                                               
#>   [8] "Peru"                                                
#>   [9] "Belgium"                                             
#>  [10] "Jersey"                                              
#>  [11] "Samoa"                                               
#>  [12] "Libya"                                               
#>  [13] "Pakistan"                                            
#>  [14] "Latvia"                                              
#>  [15] "Latvia"                                              
#>  [16] "Saudi Arabia"                                        
#>  [17] "Iceland"                                             
#>  [18] "Switzerland"                                         
#>  [19] "Vanuatu"                                             
#>  [20] "United States of America"                            
#>  [21] "Japan"                                               
#>  [22] "Yemen"                                               
#>  [23] "Cocos (Keeling) Islands"                             
#>  [24] "Philippines"                                         
#>  [25] "Macao"                                               
#>  [26] "Virgin Islands (British)"                            
#>  [27] "Estonia"                                             
#>  [28] "Congo"                                               
#>  [29] "Myanmar"                                             
#>  [30] "Guatemala"                                           
#>  [31] "Papua New Guinea"                                    
#>  [32] "Singapore"                                           
#>  [33] "Guam"                                                
#>  [34] "Haiti"                                               
#>  [35] "Virgin Islands (British)"                            
#>  [36] "Germany"                                             
#>  [37] "Bahrain"                                             
#>  [38] "Bangladesh"                                          
#>  [39] "Niger"                                               
#>  [40] "Lesotho"                                             
#>  [41] "Eritrea"                                             
#>  [42] "Slovakia"                                            
#>  [43] "Brunei Darussalam"                                   
#>  [44] "Sudan"                                               
#>  [45] "Aruba"                                               
#>  [46] "Mauritania"                                          
#>  [47] "Mayotte"                                             
#>  [48] "Czechia"                                             
#>  [49] "Morocco"                                             
#>  [50] "Barbados"                                            
#>  [51] "India"                                               
#>  [52] "Egypt"                                               
#>  [53] "Senegal"                                             
#>  [54] "Maldives"                                            
#>  [55] "Monaco"                                              
#>  [56] "Brazil"                                              
#>  [57] "Suriname"                                            
#>  [58] "Canada"                                              
#>  [59] "Cyprus"                                              
#>  [60] "Papua New Guinea"                                    
#>  [61] "Liberia"                                             
#>  [62] "Cyprus"                                              
#>  [63] "Thailand"                                            
#>  [64] "Côte d'Ivoire"                                       
#>  [65] "Eswatini"                                            
#>  [66] "Thailand"                                            
#>  [67] "New Zealand"                                         
#>  [68] "Nauru"                                               
#>  [69] "Latvia"                                              
#>  [70] "Suriname"                                            
#>  [71] "Nigeria"                                             
#>  [72] "Saint Helena, Ascension and Tristan da Cunha"        
#>  [73] "Puerto Rico"                                         
#>  [74] "Cabo Verde"                                          
#>  [75] "Bahrain"                                             
#>  [76] "Cocos (Keeling) Islands"                             
#>  [77] "Micronesia (Federated States of)"                    
#>  [78] "Afghanistan"                                         
#>  [79] "South Africa"                                        
#>  [80] "United States of America"                            
#>  [81] "Brazil"                                              
#>  [82] "Malaysia"                                            
#>  [83] "Vanuatu"                                             
#>  [84] "Seychelles"                                          
#>  [85] "Holy See"                                            
#>  [86] "Switzerland"                                         
#>  [87] "Ukraine"                                             
#>  [88] "Bhutan"                                              
#>  [89] "Syrian Arab Republic"                                
#>  [90] "Canada"                                              
#>  [91] "Uruguay"                                             
#>  [92] "Armenia"                                             
#>  [93] "China"                                               
#>  [94] "Sao Tome and Principe"                               
#>  [95] "Congo, Democratic Republic of the"                   
#>  [96] "Yemen"                                               
#>  [97] "Belarus"                                             
#>  [98] "Saint Vincent and the Grenadines"                    
#>  [99] "Slovakia"                                            
#> [100] "Cyprus"                                              
#> [101] "Slovakia"                                            
#> [102] "Qatar"                                               
#> [103] "Greenland"                                           
#> [104] "Cuba"                                                
#> [105] "Hong Kong"                                           
#> [106] "Japan"                                               
#> [107] "Congo"                                               
#> [108] "Mozambique"                                          
#> [109] "Guinea"                                              
#> [110] "Haiti"                                               
#> [111] "Tonga"                                               
#> [112] "Lithuania"                                           
#> [113] "Niger"                                               
#> [114] "Uruguay"                                             
#> [115] "Bouvet Island"                                       
#> [116] "Mali"                                                
#> [117] "Myanmar"                                             
#> [118] "Guatemala"                                           
#> [119] "Senegal"                                             
#> [120] "Bhutan"                                              
#> [121] "Poland"                                              
#> [122] "Zimbabwe"                                            
#> [123] "Greenland"                                           
#> [124] "Grenada"                                             
#> [125] "Egypt"                                               
#> [126] "Tuvalu"                                              
#> [127] "Nicaragua"                                           
#> [128] "Taiwan, Province of China"                           
#> [129] "Saint Pierre and Miquelon"                           
#> [130] "Greece"                                              
#> [131] "Bangladesh"                                          
#> [132] "Guyana"                                              
#> [133] "Nigeria"                                             
#> [134] "Bahamas"                                             
#> [135] "Slovenia"                                            
#> [136] "Ghana"                                               
#> [137] "New Zealand"                                         
#> [138] "Comoros"                                             
#> [139] "Djibouti"                                            
#> [140] "Sint Maarten (Dutch part)"                           
#> [141] "Slovenia"                                            
#> [142] "Serbia"                                              
#> [143] "Cyprus"                                              
#> [144] "Pitcairn"                                            
#> [145] "Argentina"                                           
#> [146] "Djibouti"                                            
#> [147] "Gabon"                                               
#> [148] "Bahrain"                                             
#> [149] "Peru"                                                
#> [150] "Mali"                                                
#> [151] "Cyprus"                                              
#> [152] "Brazil"                                              
#> [153] "Kyrgyzstan"                                          
#> [154] "Virgin Islands (U.S.)"                               
#> [155] "Solomon Islands"                                     
#> [156] "Cook Islands"                                        
#> [157] "Belgium"                                             
#> [158] "Tuvalu"                                              
#> [159] "Paraguay"                                            
#> [160] "Cabo Verde"                                          
#> [161] "Kyrgyzstan"                                          
#> [162] "Saint Vincent and the Grenadines"                    
#> [163] "Montserrat"                                          
#> [164] "Bermuda"                                             
#> [165] "New Caledonia"                                       
#> [166] "Angola"                                              
#> [167] "Mexico"                                              
#> [168] "United States Minor Outlying Islands"                
#> [169] "Iceland"                                             
#> [170] "Philippines"                                         
#> [171] "Ethiopia"                                            
#> [172] "Zimbabwe"                                            
#> [173] "Honduras"                                            
#> [174] "Svalbard and Jan Mayen"                              
#> [175] "Panama"                                              
#> [176] "Åland Islands"                                       
#> [177] "Lithuania"                                           
#> [178] "Mongolia"                                            
#> [179] "Tokelau"                                             
#> [180] "Sweden"                                              
#> [181] "Nigeria"                                             
#> [182] "Canada"                                              
#> [183] "Brunei Darussalam"                                   
#> [184] "Seychelles"                                          
#> [185] "Germany"                                             
#> [186] "Nepal"                                               
#> [187] "Denmark"                                             
#> [188] "Cabo Verde"                                          
#> [189] "Bangladesh"                                          
#> [190] "Honduras"                                            
#> [191] "Hungary"                                             
#> [192] "Mozambique"                                          
#> [193] "Curaçao"                                             
#> [194] "Central African Republic"                            
#> [195] "Albania"                                             
#> [196] "Barbados"                                            
#> [197] "Czechia"                                             
#> [198] "Ecuador"                                             
#> [199] "Cook Islands"                                        
#> [200] "Pakistan"                                            
#> [201] "Montenegro"                                          
#> [202] "British Indian Ocean Territory"                      
#> [203] "Sierra Leone"                                        
#> [204] "Equatorial Guinea"                                   
#> [205] "Estonia"                                             
#> [206] "Venezuela (Bolivarian Republic of)"                  
#> [207] "Greenland"                                           
#> [208] "Eritrea"                                             
#> [209] "Vanuatu"                                             
#> [210] "Brazil"                                              
#> [211] "Antigua and Barbuda"                                 
#> [212] "Brunei Darussalam"                                   
#> [213] "Mauritius"                                           
#> [214] "Sweden"                                              
#> [215] "Christmas Island"                                    
#> [216] "Comoros"                                             
#> [217] "Mauritius"                                           
#> [218] "Congo, Democratic Republic of the"                   
#> [219] "Libya"                                               
#> [220] "Lithuania"                                           
#> [221] "Solomon Islands"                                     
#> [222] "Mexico"                                              
#> [223] "Martinique"                                          
#> [224] "Greenland"                                           
#> [225] "Hong Kong"                                           
#> [226] "Tunisia"                                             
#> [227] "Pitcairn"                                            
#> [228] "Bahrain"                                             
#> [229] "Andorra"                                             
#> [230] "Sudan"                                               
#> [231] "Botswana"                                            
#> [232] "Panama"                                              
#> [233] "Mayotte"                                             
#> [234] "Libya"                                               
#> [235] "Burundi"                                             
#> [236] "Slovakia"                                            
#> [237] "Slovenia"                                            
#> [238] "Central African Republic"                            
#> [239] "Anguilla"                                            
#> [240] "Mozambique"                                          
#> [241] "Curaçao"                                             
#> [242] "Palestine, State of"                                 
#> [243] "Jordan"                                              
#> [244] "Turkey"                                              
#> [245] "Brunei Darussalam"                                   
#> [246] "Morocco"                                             
#> [247] "Virgin Islands (U.S.)"                               
#> [248] "Jamaica"                                             
#> [249] "South Sudan"                                         
#> [250] "Guernsey"                                            
#> [251] "Saudi Arabia"                                        
#> [252] "Peru"                                                
#> [253] "Åland Islands"                                       
#> [254] "Monaco"                                              
#> [255] "Albania"                                             
#> [256] "Moldova, Republic of"                                
#> [257] "Virgin Islands (British)"                            
#> [258] "Ecuador"                                             
#> [259] "Egypt"                                               
#> [260] "Côte d'Ivoire"                                       
#> [261] "Cambodia"                                            
#> [262] "Viet Nam"                                            
#> [263] "Ukraine"                                             
#> [264] "Barbados"                                            
#> [265] "Tokelau"                                             
#> [266] "Norway"                                              
#> [267] "Marshall Islands"                                    
#> [268] "Timor-Leste"                                         
#> [269] "Belize"                                              
#> [270] "Algeria"                                             
#> [271] "Uganda"                                              
#> [272] "Georgia"                                             
#> [273] "Comoros"                                             
#> [274] "Finland"                                             
#> [275] "Afghanistan"                                         
#> [276] "Saint Barthélemy"                                    
#> [277] "Saint Kitts and Nevis"                               
#> [278] "Malta"                                               
#> [279] "Cameroon"                                            
#> [280] "United States of America"                            
#> [281] "Korea, Republic of"                                  
#> [282] "Svalbard and Jan Mayen"                              
#> [283] "Eritrea"                                             
#> [284] "Mozambique"                                          
#> [285] "Zimbabwe"                                            
#> [286] "Samoa"                                               
#> [287] "Uganda"                                              
#> [288] "Saint Vincent and the Grenadines"                    
#> [289] "Mauritius"                                           
#> [290] "Mali"                                                
#> [291] "Réunion"                                             
#> [292] "Bahamas"                                             
#> [293] "Antigua and Barbuda"                                 
#> [294] "Ghana"                                               
#> [295] "Grenada"                                             
#> [296] "Belarus"                                             
#> [297] "Honduras"                                            
#> [298] "Bulgaria"                                            
#> [299] "Sao Tome and Principe"                               
#> [300] "Kazakhstan"                                          
#> [301] "Costa Rica"                                          
#> [302] "France"                                              
#> [303] "Libya"                                               
#> [304] "Angola"                                              
#> [305] "Sao Tome and Principe"                               
#> [306] "Virgin Islands (U.S.)"                               
#> [307] "Malta"                                               
#> [308] "Gabon"                                               
#> [309] "Iran (Islamic Republic of)"                          
#> [310] "Jordan"                                              
#> [311] "Åland Islands"                                       
#> [312] "Iceland"                                             
#> [313] "New Zealand"                                         
#> [314] "French Southern Territories"                         
#> [315] "Lesotho"                                             
#> [316] "Chile"                                               
#> [317] "Japan"                                               
#> [318] "Peru"                                                
#> [319] "Sudan"                                               
#> [320] "Holy See"                                            
#> [321] "Cyprus"                                              
#> [322] "Croatia"                                             
#> [323] "South Georgia and the South Sandwich Islands"        
#> [324] "Monaco"                                              
#> [325] "Maldives"                                            
#> [326] "Kuwait"                                              
#> [327] "Botswana"                                            
#> [328] "Senegal"                                             
#> [329] "Marshall Islands"                                    
#> [330] "Lebanon"                                             
#> [331] "Mayotte"                                             
#> [332] "Cuba"                                                
#> [333] "Iceland"                                             
#> [334] "Italy"                                               
#> [335] "Niue"                                                
#> [336] "Sao Tome and Principe"                               
#> [337] "Lithuania"                                           
#> [338] "Chile"                                               
#> [339] "Tanzania, United Republic of"                        
#> [340] "Colombia"                                            
#> [341] "Saint Martin (French part)"                          
#> [342] "Qatar"                                               
#> [343] "Timor-Leste"                                         
#> [344] "Canada"                                              
#> [345] "Guinea-Bissau"                                       
#> [346] "Algeria"                                             
#> [347] "Virgin Islands (U.S.)"                               
#> [348] "Papua New Guinea"                                    
#> [349] "Guinea"                                              
#> [350] "Antigua and Barbuda"                                 
#> [351] "Uzbekistan"                                          
#> [352] "Somalia"                                             
#> [353] "Virgin Islands (U.S.)"                               
#> [354] "Mozambique"                                          
#> [355] "Guadeloupe"                                          
#> [356] "Equatorial Guinea"                                   
#> [357] "Maldives"                                            
#> [358] "Armenia"                                             
#> [359] "Croatia"                                             
#> [360] "Czechia"                                             
#> [361] "Hong Kong"                                           
#> [362] "Luxembourg"                                          
#> [363] "Thailand"                                            
#> [364] "Syrian Arab Republic"                                
#> [365] "Mauritius"                                           
#> [366] "Jamaica"                                             
#> [367] "Sierra Leone"                                        
#> [368] "Tunisia"                                             
#> [369] "Estonia"                                             
#> [370] "Malaysia"                                            
#> [371] "Virgin Islands (British)"                            
#> [372] "Mexico"                                              
#> [373] "Burundi"                                             
#> [374] "Bulgaria"                                            
#> [375] "Liechtenstein"                                       
#> [376] "Seychelles"                                          
#> [377] "Vanuatu"                                             
#> [378] "Malaysia"                                            
#> [379] "Bangladesh"                                          
#> [380] "Guernsey"                                            
#> [381] "Panama"                                              
#> [382] "Anguilla"                                            
#> [383] "Uruguay"                                             
#> [384] "Poland"                                              
#> [385] "Philippines"                                         
#> [386] "Finland"                                             
#> [387] "Antarctica"                                          
#> [388] "Pitcairn"                                            
#> [389] "Montenegro"                                          
#> [390] "Russian Federation"                                  
#> [391] "Ireland"                                             
#> [392] "New Zealand"                                         
#> [393] "Niger"                                               
#> [394] "Congo, Democratic Republic of the"                   
#> [395] "Tokelau"                                             
#> [396] "Nauru"                                               
#> [397] "Ecuador"                                             
#> [398] "Sri Lanka"                                           
#> [399] "Ethiopia"                                            
#> [400] "Turkey"                                              
#> [401] "Australia"                                           
#> [402] "Seychelles"                                          
#> [403] "Réunion"                                             
#> [404] "Indonesia"                                           
#> [405] "Tokelau"                                             
#> [406] "Kenya"                                               
#> [407] "Morocco"                                             
#> [408] "Saint Pierre and Miquelon"                           
#> [409] "Indonesia"                                           
#> [410] "Mali"                                                
#> [411] "Spain"                                               
#> [412] "Iceland"                                             
#> [413] "Colombia"                                            
#> [414] "Panama"                                              
#> [415] "Heard Island and McDonald Islands"                   
#> [416] "Dominica"                                            
#> [417] "Saint Vincent and the Grenadines"                    
#> [418] "Belgium"                                             
#> [419] "Guyana"                                              
#> [420] "Ukraine"                                             
#> [421] "Algeria"                                             
#> [422] "Armenia"                                             
#> [423] "Vanuatu"                                             
#> [424] "Saint Lucia"                                         
#> [425] "Tokelau"                                             
#> [426] "Niue"                                                
#> [427] "Tokelau"                                             
#> [428] "Puerto Rico"                                         
#> [429] "Ethiopia"                                            
#> [430] "Bouvet Island"                                       
#> [431] "Mayotte"                                             
#> [432] "Bahamas"                                             
#> [433] "Sao Tome and Principe"                               
#> [434] "Japan"                                               
#> [435] "Bangladesh"                                          
#> [436] "Portugal"                                            
#> [437] "Hong Kong"                                           
#> [438] "Chile"                                               
#> [439] "Montenegro"                                          
#> [440] "Jordan"                                              
#> [441] "Sint Maarten (Dutch part)"                           
#> [442] "Antarctica"                                          
#> [443] "Macao"                                               
#> [444] "Viet Nam"                                            
#> [445] "Paraguay"                                            
#> [446] "Poland"                                              
#> [447] "Iraq"                                                
#> [448] "Viet Nam"                                            
#> [449] "Central African Republic"                            
#> [450] "Faroe Islands"                                       
#> [451] "Pitcairn"                                            
#> [452] "Samoa"                                               
#> [453] "Heard Island and McDonald Islands"                   
#> [454] "Haiti"                                               
#> [455] "Belgium"                                             
#> [456] "Andorra"                                             
#> [457] "Hungary"                                             
#> [458] "Cyprus"                                              
#> [459] "Syrian Arab Republic"                                
#> [460] "Macao"                                               
#> [461] "Uganda"                                              
#> [462] "Nepal"                                               
#> [463] "Ecuador"                                             
#> [464] "Northern Mariana Islands"                            
#> [465] "Belgium"                                             
#> [466] "United Kingdom of Great Britain and Northern Ireland"
#> [467] "Austria"                                             
#> [468] "Vanuatu"                                             
#> [469] "Guam"                                                
#> [470] "Kenya"                                               
#> [471] "Albania"                                             
#> [472] "Korea, Republic of"                                  
#> [473] "Cocos (Keeling) Islands"                             
#> [474] "British Indian Ocean Territory"                      
#> [475] "Viet Nam"                                            
#> [476] "Spain"                                               
#> [477] "Mexico"                                              
#> [478] "Malta"                                               
#> [479] "Sri Lanka"                                           
#> [480] "Antigua and Barbuda"                                 
#> [481] "Eswatini"                                            
#> [482] "China"                                               
#> [483] "Ethiopia"                                            
#> [484] "Tokelau"                                             
#> [485] "Turkey"                                              
#> [486] "Uruguay"                                             
#> [487] "Saint Pierre and Miquelon"                           
#> [488] "Curaçao"                                             
#> [489] "Niue"                                                
#> [490] "Cuba"                                                
#> [491] "Bahamas"                                             
#> [492] "Papua New Guinea"                                    
#> [493] "Virgin Islands (British)"                            
#> [494] "Gibraltar"                                           
#> [495] "Guinea-Bissau"                                       
#> [496] "Ireland"                                             
#> [497] "Saint Vincent and the Grenadines"                    
#> [498] "Syrian Arab Republic"                                
#> [499] "South Georgia and the South Sandwich Islands"        
#> [500] "Zimbabwe"                                            
random_countries(n = 5, nomenclature = "ISO3", seed = 5)
#> [1] "DJI" "IND" "RWA" "ZAF" "SVK"