Conversion table

A table containing country names in different naming conventions

Usage

country_reference_list

Format

A data frame with columns corresponding to different country naming conventions.

simple

Reference name for the geographic unit. The names in this column contain only ASCII characters. This nomenclature is available for all countries.

ISO3

3-letter country codes as defined in ISO standard 3166-1 alpha-3. This nomenclature is available for the territories in the standard (currently 249 territories).

ISO2

2-letter country codes as defined in ISO standard 3166-1 alpha-2. This nomenclature is available for the territories in the standard (currently 249 territories).

ISO_code

Numeric country codes as defined in ISO standard 3166-1 numeric. This country code is the same as the UN's country number (M49 standard). This nomenclature is available for the territories in the ISO standard (currently 249 countries).

UN_ar

Official UN name in Arabic. This nomenclature is only available for countries in the M49 standard (currently 249 territories).

UN_zh

Official UN name in Chinese. This nomenclature is only available for countries in the M49 standard (currently 249 territories).

UN_en

Official UN name in English. This nomenclature is only available for countries in the M49 standard (currently 249 territories).

UN_fr

Official UN name in French. This nomenclature is only available for countries in the M49 standard (currently 249 territories).

UN_es

Official UN name in Spanish. This nomenclature is only available for countries in the M49 standard (currently 249 territories).

UN_ru

Official UN name in Russian. This nomenclature is only available for countries in the M49 standard (currently 249 territories).

WTO_en

Official WTO name in English. This nomenclature is only available for WTO members and observers (currently 189 entities).

WTO_fr

Official WTO name in French. This nomenclature is only available for WTO members and observers (currently 189 entities).

WTO_es

Official WTO name in Spanish. This nomenclature is only available for WTO members and observers (currently 189 entities).

name_ar

Translation of ISO country names in Arabic. (currently 249 territories)

name_bg

Translation of ISO country names in Bulgarian. (currently 249 territories)

name_cs

Translation of ISO country names in Czech. (currently 249 territories)

name_da

Translation of ISO country names in Danish. (currently 249 territories)

name_de

Translation of ISO country names in German. (currently 249 territories)

name_el

Translation of ISO country names in Greek. (currently 249 territories)

name_en

Translation of ISO country names in English. (currently 249 territories)

name_es

Translation of ISO country names in Spanish. (currently 249 territories)

name_et

Translation of ISO country names in Estonian. (currently 249 territories)

name_eu

Translation of ISO country names in Basque. (currently 249 territories)

name_fi

Translation of ISO country names in Finnish. (currently 249 territories)

name_fr

Translation of ISO country names in French. (currently 249 territories)

name_hu

Translation of ISO country names in Hungarian. (currently 249 territories)

name_it

Translation of ISO country names in Italian. (currently 249 territories)

name_ja

Translation of ISO country names in Japanese. (currently 249 territories)

name_ko

Translation of ISO country names in Korean. (currently 249 territories)

name_lt

Translation of ISO country names in Lithuanian. (currently 249 territories)

name_nl

Translation of ISO country names in Dutch. (currently 249 territories)

name_no

Translation of ISO country names in Norwegian. (currently 249 territories)

name_pl

Translation of ISO country names in Polish. (currently 249 territories)

name_pt

Translation of ISO country names in Portuguese. (currently 249 territories)

name_ro

Translation of ISO country names in Romanian. (currently 249 territories)

name_ru

Translation of ISO country names in Russian. (currently 249 territories)

name_sk

Translation of ISO country names in Slovak. (currently 249 territories)

name_sv

Translation of ISO country names in Swedish. (currently 249 territories)

name_th

Translation of ISO country names in Thai. (currently 249 territories)

name_uk

Translation of ISO country names in Ukranian. (currently 249 territories)

name_zh

Translation of ISO country names in simplified Chinese. (currently 249 territories)

name_zh-tw

Translation of ISO country names in traditional Chinese. (currently 249 territories)

GTAP

GTAP country and region codes.

Name0

Other variants of the country name included to improve the matching process

Name1

Other variants of the country name included to improve the matching process

Name2

Other variants of the country name included to improve the matching process

Name3

Other variants of the country name included to improve the matching process

Name4

Other variants of the country name included to improve the matching process

Name5

Other variants of the country name included to improve the matching process

Name6

Other variants of the country name included to improve the matching process

Name7

Other variants of the country name included to improve the matching process

Name8

Other variants of the country name included to improve the matching process

Name9

Other variants of the country name included to improve the matching process

Name10

Other variants of the country name included to improve the matching process

Name11

Other variants of the country name included to improve the matching process

Name12

Other variants of the country name included to improve the matching process

Name13

Other variants of the country name included to improve the matching process

Name14

Other variants of the country name included to improve the matching process

Name15

Other variants of the country name included to improve the matching process

Name16

Other variants of the country name included to improve the matching process

Name17

Other variants of the country name included to improve the matching process

Name18

Other variants of the country name included to improve the matching process

Name19

Other variants of the country name included to improve the matching process