Output random country names

This function returns the mode of vectors. That is to say, for any given vector of values, it returns the value that appears most frequently. The function works with strings, numerical and mixed inputs. NA values are treated as distinct values.

Usage

random_countries(n, replace = FALSE, nomenclature = "name_en", seed = NULL)

Arguments

n

Number of desired (pseudo)random country names.

replace

Logical value indicating whether sampling should be with replacement.

nomenclature

Nomenclature from which the list of countries should be taken. Not all countries are present in all nomenclature, for example Taiwan is not recognised by the UN, so it will not be returned with "WTO_en". The function accept any of the nomenclatures of country_name. For a list of accepted values, refer to this page. The default is name_en, which is the English list of names in the ISO standard 3166.

seed

Single numerical value to be used as seed.

Value

A vector of n (pseudo)random country names.

See also

Examples

random_countries(10)
#>  [1] "South Sudan"                "Guam"                      
#>  [3] "Nepal"                      "Japan"                     
#>  [5] "Kiribati"                   "China"                     
#>  [7] "Bermuda"                    "Saint Martin (French part)"
#>  [9] "Marshall Islands"           "Northern Mariana Islands"  
random_countries(n = 500, replace = TRUE)
#>   [1] "Libya"                                               
#>   [2] "Botswana"                                            
#>   [3] "French Polynesia"                                    
#>   [4] "Aruba"                                               
#>   [5] "French Guiana"                                       
#>   [6] "Luxembourg"                                          
#>   [7] "France"                                              
#>   [8] "Lao People's Democratic Republic"                    
#>   [9] "Iran (Islamic Republic of)"                          
#>  [10] "Central African Republic"                            
#>  [11] "Comoros"                                             
#>  [12] "Peru"                                                
#>  [13] "Bangladesh"                                          
#>  [14] "Western Sahara"                                      
#>  [15] "Maldives"                                            
#>  [16] "Lesotho"                                             
#>  [17] "Senegal"                                             
#>  [18] "Liechtenstein"                                       
#>  [19] "Montserrat"                                          
#>  [20] "Tunisia"                                             
#>  [21] "Marshall Islands"                                    
#>  [22] "Marshall Islands"                                    
#>  [23] "Liechtenstein"                                       
#>  [24] "Micronesia (Federated States of)"                    
#>  [25] "Estonia"                                             
#>  [26] "South Georgia and the South Sandwich Islands"        
#>  [27] "Belize"                                              
#>  [28] "Syrian Arab Republic"                                
#>  [29] "Hungary"                                             
#>  [30] "Nepal"                                               
#>  [31] "Ecuador"                                             
#>  [32] "Montenegro"                                          
#>  [33] "Guadeloupe"                                          
#>  [34] "Montserrat"                                          
#>  [35] "Cameroon"                                            
#>  [36] "Czechia"                                             
#>  [37] "Cyprus"                                              
#>  [38] "Mongolia"                                            
#>  [39] "Tonga"                                               
#>  [40] "Solomon Islands"                                     
#>  [41] "Mozambique"                                          
#>  [42] "Dominica"                                            
#>  [43] "Kenya"                                               
#>  [44] "Israel"                                              
#>  [45] "Bulgaria"                                            
#>  [46] "Yemen"                                               
#>  [47] "Holy See"                                            
#>  [48] "Heard Island and McDonald Islands"                   
#>  [49] "Belarus"                                             
#>  [50] "Turkey"                                              
#>  [51] "Switzerland"                                         
#>  [52] "Netherlands"                                         
#>  [53] "Saint Barthélemy"                                    
#>  [54] "Dominica"                                            
#>  [55] "Barbados"                                            
#>  [56] "Bulgaria"                                            
#>  [57] "Austria"                                             
#>  [58] "Saint Helena, Ascension and Tristan da Cunha"        
#>  [59] "Thailand"                                            
#>  [60] "Mozambique"                                          
#>  [61] "Bolivia (Plurinational State of)"                    
#>  [62] "Cameroon"                                            
#>  [63] "Macao"                                               
#>  [64] "Holy See"                                            
#>  [65] "Gabon"                                               
#>  [66] "Montenegro"                                          
#>  [67] "Burkina Faso"                                        
#>  [68] "Paraguay"                                            
#>  [69] "Uruguay"                                             
#>  [70] "French Guiana"                                       
#>  [71] "Germany"                                             
#>  [72] "Faroe Islands"                                       
#>  [73] "South Africa"                                        
#>  [74] "Dominica"                                            
#>  [75] "Saudi Arabia"                                        
#>  [76] "Paraguay"                                            
#>  [77] "Nigeria"                                             
#>  [78] "Bonaire, Sint Eustatius and Saba"                    
#>  [79] "Bahrain"                                             
#>  [80] "Seychelles"                                          
#>  [81] "Virgin Islands (U.S.)"                               
#>  [82] "Italy"                                               
#>  [83] "Tunisia"                                             
#>  [84] "Haiti"                                               
#>  [85] "Heard Island and McDonald Islands"                   
#>  [86] "Saint Barthélemy"                                    
#>  [87] "Senegal"                                             
#>  [88] "Saint Helena, Ascension and Tristan da Cunha"        
#>  [89] "Luxembourg"                                          
#>  [90] "Comoros"                                             
#>  [91] "Namibia"                                             
#>  [92] "Uganda"                                              
#>  [93] "Madagascar"                                          
#>  [94] "Lithuania"                                           
#>  [95] "Bhutan"                                              
#>  [96] "Guinea"                                              
#>  [97] "American Samoa"                                      
#>  [98] "Togo"                                                
#>  [99] "Cabo Verde"                                          
#> [100] "Hungary"                                             
#> [101] "Tokelau"                                             
#> [102] "Djibouti"                                            
#> [103] "Georgia"                                             
#> [104] "Andorra"                                             
#> [105] "Panama"                                              
#> [106] "Bosnia and Herzegovina"                              
#> [107] "Spain"                                               
#> [108] "South Georgia and the South Sandwich Islands"        
#> [109] "Jersey"                                              
#> [110] "Saint Helena, Ascension and Tristan da Cunha"        
#> [111] "Spain"                                               
#> [112] "Namibia"                                             
#> [113] "South Georgia and the South Sandwich Islands"        
#> [114] "Micronesia (Federated States of)"                    
#> [115] "Syrian Arab Republic"                                
#> [116] "Pitcairn"                                            
#> [117] "Mozambique"                                          
#> [118] "Bahrain"                                             
#> [119] "Venezuela (Bolivarian Republic of)"                  
#> [120] "Ethiopia"                                            
#> [121] "Portugal"                                            
#> [122] "Saint Helena, Ascension and Tristan da Cunha"        
#> [123] "Bhutan"                                              
#> [124] "Greenland"                                           
#> [125] "Slovenia"                                            
#> [126] "Micronesia (Federated States of)"                    
#> [127] "Romania"                                             
#> [128] "Aruba"                                               
#> [129] "Iran (Islamic Republic of)"                          
#> [130] "Saint Kitts and Nevis"                               
#> [131] "Estonia"                                             
#> [132] "Curaçao"                                             
#> [133] "Libya"                                               
#> [134] "Armenia"                                             
#> [135] "Central African Republic"                            
#> [136] "Virgin Islands (British)"                            
#> [137] "Bulgaria"                                            
#> [138] "Bhutan"                                              
#> [139] "Myanmar"                                             
#> [140] "Nicaragua"                                           
#> [141] "Panama"                                              
#> [142] "Greece"                                              
#> [143] "Peru"                                                
#> [144] "Saint Helena, Ascension and Tristan da Cunha"        
#> [145] "Svalbard and Jan Mayen"                              
#> [146] "Estonia"                                             
#> [147] "Zimbabwe"                                            
#> [148] "Estonia"                                             
#> [149] "South Georgia and the South Sandwich Islands"        
#> [150] "Hong Kong"                                           
#> [151] "Jordan"                                              
#> [152] "United States of America"                            
#> [153] "Qatar"                                               
#> [154] "Liberia"                                             
#> [155] "Côte d'Ivoire"                                       
#> [156] "Bulgaria"                                            
#> [157] "Namibia"                                             
#> [158] "Namibia"                                             
#> [159] "Togo"                                                
#> [160] "Libya"                                               
#> [161] "Peru"                                                
#> [162] "Palau"                                               
#> [163] "Cabo Verde"                                          
#> [164] "Antarctica"                                          
#> [165] "Slovenia"                                            
#> [166] "Antarctica"                                          
#> [167] "Solomon Islands"                                     
#> [168] "Montenegro"                                          
#> [169] "Tokelau"                                             
#> [170] "United Arab Emirates"                                
#> [171] "Seychelles"                                          
#> [172] "Heard Island and McDonald Islands"                   
#> [173] "Colombia"                                            
#> [174] "Sint Maarten (Dutch part)"                           
#> [175] "Suriname"                                            
#> [176] "Gibraltar"                                           
#> [177] "Uruguay"                                             
#> [178] "Korea (Democratic People's Republic of)"             
#> [179] "Åland Islands"                                       
#> [180] "Costa Rica"                                          
#> [181] "Ghana"                                               
#> [182] "American Samoa"                                      
#> [183] "Moldova, Republic of"                                
#> [184] "Italy"                                               
#> [185] "Turkey"                                              
#> [186] "Vanuatu"                                             
#> [187] "Taiwan, Province of China"                           
#> [188] "Togo"                                                
#> [189] "Liechtenstein"                                       
#> [190] "Viet Nam"                                            
#> [191] "Italy"                                               
#> [192] "Morocco"                                             
#> [193] "Monaco"                                              
#> [194] "Paraguay"                                            
#> [195] "Peru"                                                
#> [196] "Central African Republic"                            
#> [197] "Algeria"                                             
#> [198] "Tonga"                                               
#> [199] "San Marino"                                          
#> [200] "Armenia"                                             
#> [201] "Albania"                                             
#> [202] "Côte d'Ivoire"                                       
#> [203] "Mauritius"                                           
#> [204] "Netherlands"                                         
#> [205] "Macao"                                               
#> [206] "Saint Helena, Ascension and Tristan da Cunha"        
#> [207] "Georgia"                                             
#> [208] "Lao People's Democratic Republic"                    
#> [209] "Saint Martin (French part)"                          
#> [210] "Monaco"                                              
#> [211] "Sint Maarten (Dutch part)"                           
#> [212] "Puerto Rico"                                         
#> [213] "Congo"                                               
#> [214] "Norway"                                              
#> [215] "Mauritius"                                           
#> [216] "Costa Rica"                                          
#> [217] "Réunion"                                             
#> [218] "Saint Lucia"                                         
#> [219] "Monaco"                                              
#> [220] "Slovenia"                                            
#> [221] "Belarus"                                             
#> [222] "Finland"                                             
#> [223] "Japan"                                               
#> [224] "Saint Kitts and Nevis"                               
#> [225] "Moldova, Republic of"                                
#> [226] "French Southern Territories"                         
#> [227] "Madagascar"                                          
#> [228] "British Indian Ocean Territory"                      
#> [229] "El Salvador"                                         
#> [230] "Honduras"                                            
#> [231] "El Salvador"                                         
#> [232] "Guinea"                                              
#> [233] "Mayotte"                                             
#> [234] "Saint Kitts and Nevis"                               
#> [235] "Turkmenistan"                                        
#> [236] "Canada"                                              
#> [237] "Gibraltar"                                           
#> [238] "Korea (Democratic People's Republic of)"             
#> [239] "Lithuania"                                           
#> [240] "Moldova, Republic of"                                
#> [241] "Cayman Islands"                                      
#> [242] "Estonia"                                             
#> [243] "Zimbabwe"                                            
#> [244] "Luxembourg"                                          
#> [245] "Haiti"                                               
#> [246] "Slovenia"                                            
#> [247] "Jamaica"                                             
#> [248] "Bhutan"                                              
#> [249] "India"                                               
#> [250] "Anguilla"                                            
#> [251] "Suriname"                                            
#> [252] "Andorra"                                             
#> [253] "Antigua and Barbuda"                                 
#> [254] "Armenia"                                             
#> [255] "Andorra"                                             
#> [256] "Cuba"                                                
#> [257] "Tonga"                                               
#> [258] "Serbia"                                              
#> [259] "Aruba"                                               
#> [260] "Belgium"                                             
#> [261] "Guinea"                                              
#> [262] "Kyrgyzstan"                                          
#> [263] "Guatemala"                                           
#> [264] "Serbia"                                              
#> [265] "Denmark"                                             
#> [266] "Moldova, Republic of"                                
#> [267] "Greenland"                                           
#> [268] "Cocos (Keeling) Islands"                             
#> [269] "San Marino"                                          
#> [270] "Cayman Islands"                                      
#> [271] "Turks and Caicos Islands"                            
#> [272] "Mongolia"                                            
#> [273] "Latvia"                                              
#> [274] "Ecuador"                                             
#> [275] "Singapore"                                           
#> [276] "Mauritania"                                          
#> [277] "French Polynesia"                                    
#> [278] "Pitcairn"                                            
#> [279] "Zimbabwe"                                            
#> [280] "Italy"                                               
#> [281] "Niue"                                                
#> [282] "Nauru"                                               
#> [283] "Guernsey"                                            
#> [284] "Afghanistan"                                         
#> [285] "Japan"                                               
#> [286] "Pitcairn"                                            
#> [287] "Afghanistan"                                         
#> [288] "Solomon Islands"                                     
#> [289] "Tajikistan"                                          
#> [290] "Barbados"                                            
#> [291] "Anguilla"                                            
#> [292] "Jordan"                                              
#> [293] "Tunisia"                                             
#> [294] "Kiribati"                                            
#> [295] "Mozambique"                                          
#> [296] "United Arab Emirates"                                
#> [297] "Mauritania"                                          
#> [298] "Trinidad and Tobago"                                 
#> [299] "Moldova, Republic of"                                
#> [300] "Anguilla"                                            
#> [301] "Romania"                                             
#> [302] "United Arab Emirates"                                
#> [303] "Liechtenstein"                                       
#> [304] "Micronesia (Federated States of)"                    
#> [305] "Chile"                                               
#> [306] "Bermuda"                                             
#> [307] "Guernsey"                                            
#> [308] "Antigua and Barbuda"                                 
#> [309] "Antarctica"                                          
#> [310] "Gibraltar"                                           
#> [311] "Mongolia"                                            
#> [312] "Tokelau"                                             
#> [313] "Solomon Islands"                                     
#> [314] "British Indian Ocean Territory"                      
#> [315] "China"                                               
#> [316] "Senegal"                                             
#> [317] "Réunion"                                             
#> [318] "Burkina Faso"                                        
#> [319] "Hong Kong"                                           
#> [320] "Nicaragua"                                           
#> [321] "Liechtenstein"                                       
#> [322] "Papua New Guinea"                                    
#> [323] "Tanzania, United Republic of"                        
#> [324] "Rwanda"                                              
#> [325] "Wallis and Futuna"                                   
#> [326] "Saint Barthélemy"                                    
#> [327] "Aruba"                                               
#> [328] "Morocco"                                             
#> [329] "Argentina"                                           
#> [330] "Seychelles"                                          
#> [331] "Guinea-Bissau"                                       
#> [332] "Grenada"                                             
#> [333] "Sierra Leone"                                        
#> [334] "Saint Pierre and Miquelon"                           
#> [335] "Djibouti"                                            
#> [336] "Liberia"                                             
#> [337] "Viet Nam"                                            
#> [338] "Albania"                                             
#> [339] "Pakistan"                                            
#> [340] "Venezuela (Bolivarian Republic of)"                  
#> [341] "Iraq"                                                
#> [342] "Saint Lucia"                                         
#> [343] "Australia"                                           
#> [344] "Sweden"                                              
#> [345] "Thailand"                                            
#> [346] "American Samoa"                                      
#> [347] "Bulgaria"                                            
#> [348] "Bhutan"                                              
#> [349] "Rwanda"                                              
#> [350] "Moldova, Republic of"                                
#> [351] "Montenegro"                                          
#> [352] "Mexico"                                              
#> [353] "Oman"                                                
#> [354] "India"                                               
#> [355] "Georgia"                                             
#> [356] "Bahrain"                                             
#> [357] "Turks and Caicos Islands"                            
#> [358] "Brazil"                                              
#> [359] "China"                                               
#> [360] "Portugal"                                            
#> [361] "Congo"                                               
#> [362] "Thailand"                                            
#> [363] "Macao"                                               
#> [364] "Virgin Islands (British)"                            
#> [365] "Zimbabwe"                                            
#> [366] "Iceland"                                             
#> [367] "Liberia"                                             
#> [368] "Belize"                                              
#> [369] "Denmark"                                             
#> [370] "Montserrat"                                          
#> [371] "Bolivia (Plurinational State of)"                    
#> [372] "Cambodia"                                            
#> [373] "Belgium"                                             
#> [374] "United States Minor Outlying Islands"                
#> [375] "Iran (Islamic Republic of)"                          
#> [376] "Gibraltar"                                           
#> [377] "South Africa"                                        
#> [378] "Puerto Rico"                                         
#> [379] "Oman"                                                
#> [380] "Brazil"                                              
#> [381] "Romania"                                             
#> [382] "New Caledonia"                                       
#> [383] "Guyana"                                              
#> [384] "Burundi"                                             
#> [385] "Cayman Islands"                                      
#> [386] "Korea, Republic of"                                  
#> [387] "Iran (Islamic Republic of)"                          
#> [388] "Uzbekistan"                                          
#> [389] "Belarus"                                             
#> [390] "Suriname"                                            
#> [391] "Mongolia"                                            
#> [392] "Martinique"                                          
#> [393] "Yemen"                                               
#> [394] "Saint Barthélemy"                                    
#> [395] "Sweden"                                              
#> [396] "Romania"                                             
#> [397] "Mauritius"                                           
#> [398] "Svalbard and Jan Mayen"                              
#> [399] "Angola"                                              
#> [400] "Estonia"                                             
#> [401] "Uganda"                                              
#> [402] "Gabon"                                               
#> [403] "Mongolia"                                            
#> [404] "Algeria"                                             
#> [405] "Congo"                                               
#> [406] "Senegal"                                             
#> [407] "Mali"                                                
#> [408] "Niger"                                               
#> [409] "British Indian Ocean Territory"                      
#> [410] "Cambodia"                                            
#> [411] "Zimbabwe"                                            
#> [412] "Aruba"                                               
#> [413] "Peru"                                                
#> [414] "Saint Pierre and Miquelon"                           
#> [415] "Azerbaijan"                                          
#> [416] "Senegal"                                             
#> [417] "Croatia"                                             
#> [418] "Sao Tome and Principe"                               
#> [419] "Saint Lucia"                                         
#> [420] "Cyprus"                                              
#> [421] "Honduras"                                            
#> [422] "Guadeloupe"                                          
#> [423] "Chad"                                                
#> [424] "Grenada"                                             
#> [425] "Macao"                                               
#> [426] "Syrian Arab Republic"                                
#> [427] "Spain"                                               
#> [428] "Portugal"                                            
#> [429] "Grenada"                                             
#> [430] "Lao People's Democratic Republic"                    
#> [431] "Christmas Island"                                    
#> [432] "Saint Lucia"                                         
#> [433] "Barbados"                                            
#> [434] "China"                                               
#> [435] "Montenegro"                                          
#> [436] "Bahamas"                                             
#> [437] "Guinea"                                              
#> [438] "Honduras"                                            
#> [439] "Mauritania"                                          
#> [440] "Georgia"                                             
#> [441] "Mozambique"                                          
#> [442] "Togo"                                                
#> [443] "Gabon"                                               
#> [444] "Pitcairn"                                            
#> [445] "United Arab Emirates"                                
#> [446] "Rwanda"                                              
#> [447] "United States Minor Outlying Islands"                
#> [448] "Mongolia"                                            
#> [449] "Ghana"                                               
#> [450] "Colombia"                                            
#> [451] "Malta"                                               
#> [452] "Bulgaria"                                            
#> [453] "Sweden"                                              
#> [454] "Mongolia"                                            
#> [455] "Togo"                                                
#> [456] "Spain"                                               
#> [457] "Israel"                                              
#> [458] "Saint Helena, Ascension and Tristan da Cunha"        
#> [459] "Cameroon"                                            
#> [460] "Antigua and Barbuda"                                 
#> [461] "Cayman Islands"                                      
#> [462] "Somalia"                                             
#> [463] "Saint Pierre and Miquelon"                           
#> [464] "Denmark"                                             
#> [465] "Jamaica"                                             
#> [466] "Cocos (Keeling) Islands"                             
#> [467] "United Kingdom of Great Britain and Northern Ireland"
#> [468] "Syrian Arab Republic"                                
#> [469] "Micronesia (Federated States of)"                    
#> [470] "Comoros"                                             
#> [471] "Iraq"                                                
#> [472] "Mayotte"                                             
#> [473] "Burkina Faso"                                        
#> [474] "Antarctica"                                          
#> [475] "Netherlands"                                         
#> [476] "Dominican Republic"                                  
#> [477] "Turkey"                                              
#> [478] "United States Minor Outlying Islands"                
#> [479] "Sint Maarten (Dutch part)"                           
#> [480] "French Guiana"                                       
#> [481] "Dominican Republic"                                  
#> [482] "Congo, Democratic Republic of the"                   
#> [483] "Ethiopia"                                            
#> [484] "Ireland"                                             
#> [485] "Czechia"                                             
#> [486] "Tunisia"                                             
#> [487] "Denmark"                                             
#> [488] "Fiji"                                                
#> [489] "Lithuania"                                           
#> [490] "Anguilla"                                            
#> [491] "Eritrea"                                             
#> [492] "Barbados"                                            
#> [493] "Thailand"                                            
#> [494] "Gibraltar"                                           
#> [495] "Liberia"                                             
#> [496] "Wallis and Futuna"                                   
#> [497] "Montserrat"                                          
#> [498] "Comoros"                                             
#> [499] "Greenland"                                           
#> [500] "Tuvalu"                                              
random_countries(n = 5, nomenclature = "ISO3", seed = 5)
#> [1] "DJI" "IND" "RWA" "ZAF" "SVK"