Output random country names

This function returns the mode of vectors. That is to say, for any given vector of values, it returns the value that appears most frequently. The function works with strings, numerical and mixed inputs. NA values are treated as distinct values.

Usage

random_countries(n, replace = FALSE, nomenclature = "name_en", seed = NULL)

Arguments

n

Number of desired (pseudo)random country names.

replace

Logical value indicating whether sampling should be with replacement.

nomenclature

Nomenclature from which the list of countries should be taken. Not all countries are present in all nomenclature, for example Taiwan is not recognised by the UN, so it will not be returned with "WTO_en". The function accept any of the nomenclatures of country_name. For a list of accepted values, refer to this page. The default is name_en, which is the English list of names in the ISO standard 3166.

seed

Single numerical value to be used as seed.

Value

A vector of n (pseudo)random country names.

See also

Examples

random_countries(10)
#>  [1] "Gabon"               "Eswatini"            "Seychelles"         
#>  [4] "Belgium"             "Mexico"              "Madagascar"         
#>  [7] "Anguilla"            "Palestine, State of" "Algeria"            
#> [10] "Cameroon"           
random_countries(n = 500, replace = TRUE)
#>   [1] "Liberia"                                             
#>   [2] "Marshall Islands"                                    
#>   [3] "Singapore"                                           
#>   [4] "Morocco"                                             
#>   [5] "Italy"                                               
#>   [6] "Lesotho"                                             
#>   [7] "Faroe Islands"                                       
#>   [8] "Montenegro"                                          
#>   [9] "Fiji"                                                
#>  [10] "Jordan"                                              
#>  [11] "Kyrgyzstan"                                          
#>  [12] "Palestine, State of"                                 
#>  [13] "Iraq"                                                
#>  [14] "San Marino"                                          
#>  [15] "Yemen"                                               
#>  [16] "Holy See"                                            
#>  [17] "Central African Republic"                            
#>  [18] "Netherlands"                                         
#>  [19] "Canada"                                              
#>  [20] "Burkina Faso"                                        
#>  [21] "Greece"                                              
#>  [22] "Malaysia"                                            
#>  [23] "Uganda"                                              
#>  [24] "Armenia"                                             
#>  [25] "Anguilla"                                            
#>  [26] "Sao Tome and Principe"                               
#>  [27] "Jamaica"                                             
#>  [28] "Sierra Leone"                                        
#>  [29] "Maldives"                                            
#>  [30] "El Salvador"                                         
#>  [31] "Palestine, State of"                                 
#>  [32] "Palau"                                               
#>  [33] "Bhutan"                                              
#>  [34] "Costa Rica"                                          
#>  [35] "Togo"                                                
#>  [36] "Cuba"                                                
#>  [37] "Ethiopia"                                            
#>  [38] "Mauritius"                                           
#>  [39] "Réunion"                                             
#>  [40] "Paraguay"                                            
#>  [41] "Christmas Island"                                    
#>  [42] "Zambia"                                              
#>  [43] "Niue"                                                
#>  [44] "Romania"                                             
#>  [45] "Faroe Islands"                                       
#>  [46] "Eswatini"                                            
#>  [47] "Burkina Faso"                                        
#>  [48] "Saint Helena, Ascension and Tristan da Cunha"        
#>  [49] "Holy See"                                            
#>  [50] "Jamaica"                                             
#>  [51] "Saudi Arabia"                                        
#>  [52] "Monaco"                                              
#>  [53] "Bangladesh"                                          
#>  [54] "Burkina Faso"                                        
#>  [55] "Mali"                                                
#>  [56] "Solomon Islands"                                     
#>  [57] "Svalbard and Jan Mayen"                              
#>  [58] "Tuvalu"                                              
#>  [59] "Aruba"                                               
#>  [60] "Saint Kitts and Nevis"                               
#>  [61] "Cocos (Keeling) Islands"                             
#>  [62] "Saint Pierre and Miquelon"                           
#>  [63] "Comoros"                                             
#>  [64] "Zambia"                                              
#>  [65] "Martinique"                                          
#>  [66] "El Salvador"                                         
#>  [67] "Uruguay"                                             
#>  [68] "Iceland"                                             
#>  [69] "Iceland"                                             
#>  [70] "Grenada"                                             
#>  [71] "Guadeloupe"                                          
#>  [72] "Somalia"                                             
#>  [73] "Cook Islands"                                        
#>  [74] "United States Minor Outlying Islands"                
#>  [75] "Bermuda"                                             
#>  [76] "Burkina Faso"                                        
#>  [77] "Réunion"                                             
#>  [78] "Falkland Islands (Malvinas)"                         
#>  [79] "Denmark"                                             
#>  [80] "Portugal"                                            
#>  [81] "Turkey"                                              
#>  [82] "Iceland"                                             
#>  [83] "Canada"                                              
#>  [84] "Madagascar"                                          
#>  [85] "Bonaire, Sint Eustatius and Saba"                    
#>  [86] "Tokelau"                                             
#>  [87] "Senegal"                                             
#>  [88] "Malawi"                                              
#>  [89] "India"                                               
#>  [90] "New Zealand"                                         
#>  [91] "Czechia"                                             
#>  [92] "Gambia"                                              
#>  [93] "Croatia"                                             
#>  [94] "Yemen"                                               
#>  [95] "Samoa"                                               
#>  [96] "Guinea"                                              
#>  [97] "Luxembourg"                                          
#>  [98] "Venezuela (Bolivarian Republic of)"                  
#>  [99] "Mauritius"                                           
#> [100] "Egypt"                                               
#> [101] "Panama"                                              
#> [102] "American Samoa"                                      
#> [103] "Congo, Democratic Republic of the"                   
#> [104] "Singapore"                                           
#> [105] "Cyprus"                                              
#> [106] "Liechtenstein"                                       
#> [107] "Kuwait"                                              
#> [108] "Costa Rica"                                          
#> [109] "Egypt"                                               
#> [110] "United Kingdom of Great Britain and Northern Ireland"
#> [111] "South Africa"                                        
#> [112] "Samoa"                                               
#> [113] "Japan"                                               
#> [114] "Ghana"                                               
#> [115] "Cameroon"                                            
#> [116] "Sao Tome and Principe"                               
#> [117] "Bahamas"                                             
#> [118] "Solomon Islands"                                     
#> [119] "Russian Federation"                                  
#> [120] "Poland"                                              
#> [121] "Nepal"                                               
#> [122] "Malta"                                               
#> [123] "South Sudan"                                         
#> [124] "Sierra Leone"                                        
#> [125] "Iran (Islamic Republic of)"                          
#> [126] "Timor-Leste"                                         
#> [127] "Eswatini"                                            
#> [128] "Djibouti"                                            
#> [129] "Kenya"                                               
#> [130] "Gabon"                                               
#> [131] "Sint Maarten (Dutch part)"                           
#> [132] "Andorra"                                             
#> [133] "United Kingdom of Great Britain and Northern Ireland"
#> [134] "Eritrea"                                             
#> [135] "Turks and Caicos Islands"                            
#> [136] "Liberia"                                             
#> [137] "Cyprus"                                              
#> [138] "Spain"                                               
#> [139] "Puerto Rico"                                         
#> [140] "Bulgaria"                                            
#> [141] "Liberia"                                             
#> [142] "Iceland"                                             
#> [143] "Thailand"                                            
#> [144] "Palau"                                               
#> [145] "Bosnia and Herzegovina"                              
#> [146] "Guyana"                                              
#> [147] "Saint Vincent and the Grenadines"                    
#> [148] "Czechia"                                             
#> [149] "Isle of Man"                                         
#> [150] "Macao"                                               
#> [151] "Norway"                                              
#> [152] "Montenegro"                                          
#> [153] "Guernsey"                                            
#> [154] "South Africa"                                        
#> [155] "Somalia"                                             
#> [156] "Equatorial Guinea"                                   
#> [157] "Gibraltar"                                           
#> [158] "Singapore"                                           
#> [159] "Christmas Island"                                    
#> [160] "Congo, Democratic Republic of the"                   
#> [161] "Bonaire, Sint Eustatius and Saba"                    
#> [162] "Eritrea"                                             
#> [163] "Luxembourg"                                          
#> [164] "Bahamas"                                             
#> [165] "Seychelles"                                          
#> [166] "France"                                              
#> [167] "Montserrat"                                          
#> [168] "Albania"                                             
#> [169] "Nigeria"                                             
#> [170] "Faroe Islands"                                       
#> [171] "Saint Vincent and the Grenadines"                    
#> [172] "Palau"                                               
#> [173] "Qatar"                                               
#> [174] "Cyprus"                                              
#> [175] "Guatemala"                                           
#> [176] "Yemen"                                               
#> [177] "Equatorial Guinea"                                   
#> [178] "Indonesia"                                           
#> [179] "Sri Lanka"                                           
#> [180] "Azerbaijan"                                          
#> [181] "Estonia"                                             
#> [182] "Hungary"                                             
#> [183] "United Kingdom of Great Britain and Northern Ireland"
#> [184] "Saint Lucia"                                         
#> [185] "Pakistan"                                            
#> [186] "Bonaire, Sint Eustatius and Saba"                    
#> [187] "Macao"                                               
#> [188] "Haiti"                                               
#> [189] "Azerbaijan"                                          
#> [190] "Kenya"                                               
#> [191] "Bermuda"                                             
#> [192] "Algeria"                                             
#> [193] "Liberia"                                             
#> [194] "Kyrgyzstan"                                          
#> [195] "Palestine, State of"                                 
#> [196] "Iran (Islamic Republic of)"                          
#> [197] "Barbados"                                            
#> [198] "Sri Lanka"                                           
#> [199] "Saint Pierre and Miquelon"                           
#> [200] "Madagascar"                                          
#> [201] "Falkland Islands (Malvinas)"                         
#> [202] "Bonaire, Sint Eustatius and Saba"                    
#> [203] "Colombia"                                            
#> [204] "Montserrat"                                          
#> [205] "France"                                              
#> [206] "Zambia"                                              
#> [207] "Austria"                                             
#> [208] "South Georgia and the South Sandwich Islands"        
#> [209] "North Macedonia"                                     
#> [210] "Maldives"                                            
#> [211] "Croatia"                                             
#> [212] "Israel"                                              
#> [213] "Antigua and Barbuda"                                 
#> [214] "Poland"                                              
#> [215] "Dominican Republic"                                  
#> [216] "United Kingdom of Great Britain and Northern Ireland"
#> [217] "Anguilla"                                            
#> [218] "South Georgia and the South Sandwich Islands"        
#> [219] "Argentina"                                           
#> [220] "Palestine, State of"                                 
#> [221] "Gambia"                                              
#> [222] "Bahrain"                                             
#> [223] "Greenland"                                           
#> [224] "Cayman Islands"                                      
#> [225] "Honduras"                                            
#> [226] "Solomon Islands"                                     
#> [227] "Sudan"                                               
#> [228] "Sierra Leone"                                        
#> [229] "Kenya"                                               
#> [230] "Moldova, Republic of"                                
#> [231] "Slovenia"                                            
#> [232] "Saint Helena, Ascension and Tristan da Cunha"        
#> [233] "Malawi"                                              
#> [234] "Tajikistan"                                          
#> [235] "Western Sahara"                                      
#> [236] "Papua New Guinea"                                    
#> [237] "Guatemala"                                           
#> [238] "Saudi Arabia"                                        
#> [239] "Colombia"                                            
#> [240] "Faroe Islands"                                       
#> [241] "Panama"                                              
#> [242] "Honduras"                                            
#> [243] "Moldova, Republic of"                                
#> [244] "Dominican Republic"                                  
#> [245] "Grenada"                                             
#> [246] "Uzbekistan"                                          
#> [247] "Bouvet Island"                                       
#> [248] "United Arab Emirates"                                
#> [249] "Palau"                                               
#> [250] "Netherlands"                                         
#> [251] "Argentina"                                           
#> [252] "Andorra"                                             
#> [253] "Barbados"                                            
#> [254] "Croatia"                                             
#> [255] "Mozambique"                                          
#> [256] "Saudi Arabia"                                        
#> [257] "Cyprus"                                              
#> [258] "Moldova, Republic of"                                
#> [259] "Cambodia"                                            
#> [260] "Fiji"                                                
#> [261] "Somalia"                                             
#> [262] "Sweden"                                              
#> [263] "Comoros"                                             
#> [264] "Kazakhstan"                                          
#> [265] "Lithuania"                                           
#> [266] "Aruba"                                               
#> [267] "Equatorial Guinea"                                   
#> [268] "Italy"                                               
#> [269] "Central African Republic"                            
#> [270] "Cameroon"                                            
#> [271] "Togo"                                                
#> [272] "Saint Kitts and Nevis"                               
#> [273] "Belgium"                                             
#> [274] "El Salvador"                                         
#> [275] "Yemen"                                               
#> [276] "French Guiana"                                       
#> [277] "Nepal"                                               
#> [278] "Mauritania"                                          
#> [279] "China"                                               
#> [280] "Fiji"                                                
#> [281] "Saint Barthélemy"                                    
#> [282] "Mali"                                                
#> [283] "Mayotte"                                             
#> [284] "Pitcairn"                                            
#> [285] "Palestine, State of"                                 
#> [286] "Seychelles"                                          
#> [287] "Afghanistan"                                         
#> [288] "French Southern Territories"                         
#> [289] "Puerto Rico"                                         
#> [290] "Guam"                                                
#> [291] "Trinidad and Tobago"                                 
#> [292] "Ghana"                                               
#> [293] "South Africa"                                        
#> [294] "Faroe Islands"                                       
#> [295] "Sao Tome and Principe"                               
#> [296] "Jordan"                                              
#> [297] "Pakistan"                                            
#> [298] "Kuwait"                                              
#> [299] "Lesotho"                                             
#> [300] "Virgin Islands (U.S.)"                               
#> [301] "Northern Mariana Islands"                            
#> [302] "Panama"                                              
#> [303] "Iraq"                                                
#> [304] "French Southern Territories"                         
#> [305] "Angola"                                              
#> [306] "Côte d'Ivoire"                                       
#> [307] "Kyrgyzstan"                                          
#> [308] "Benin"                                               
#> [309] "Brunei Darussalam"                                   
#> [310] "American Samoa"                                      
#> [311] "Micronesia (Federated States of)"                    
#> [312] "Uzbekistan"                                          
#> [313] "Sri Lanka"                                           
#> [314] "Haiti"                                               
#> [315] "Suriname"                                            
#> [316] "Micronesia (Federated States of)"                    
#> [317] "United Kingdom of Great Britain and Northern Ireland"
#> [318] "Bahamas"                                             
#> [319] "Sint Maarten (Dutch part)"                           
#> [320] "Japan"                                               
#> [321] "Turks and Caicos Islands"                            
#> [322] "Maldives"                                            
#> [323] "Spain"                                               
#> [324] "Guyana"                                              
#> [325] "Tanzania, United Republic of"                        
#> [326] "Uruguay"                                             
#> [327] "Barbados"                                            
#> [328] "Azerbaijan"                                          
#> [329] "Sri Lanka"                                           
#> [330] "Guam"                                                
#> [331] "Nicaragua"                                           
#> [332] "Wallis and Futuna"                                   
#> [333] "Guam"                                                
#> [334] "Mali"                                                
#> [335] "France"                                              
#> [336] "Western Sahara"                                      
#> [337] "Somalia"                                             
#> [338] "Eswatini"                                            
#> [339] "Croatia"                                             
#> [340] "American Samoa"                                      
#> [341] "Belgium"                                             
#> [342] "Greenland"                                           
#> [343] "Kyrgyzstan"                                          
#> [344] "South Sudan"                                         
#> [345] "Comoros"                                             
#> [346] "Korea, Republic of"                                  
#> [347] "New Caledonia"                                       
#> [348] "Trinidad and Tobago"                                 
#> [349] "Zimbabwe"                                            
#> [350] "Marshall Islands"                                    
#> [351] "Saint Martin (French part)"                          
#> [352] "New Zealand"                                         
#> [353] "Holy See"                                            
#> [354] "Holy See"                                            
#> [355] "Curaçao"                                             
#> [356] "American Samoa"                                      
#> [357] "Kyrgyzstan"                                          
#> [358] "Thailand"                                            
#> [359] "Vanuatu"                                             
#> [360] "Tokelau"                                             
#> [361] "Tonga"                                               
#> [362] "Viet Nam"                                            
#> [363] "Indonesia"                                           
#> [364] "Samoa"                                               
#> [365] "Egypt"                                               
#> [366] "Poland"                                              
#> [367] "Saint Kitts and Nevis"                               
#> [368] "Brunei Darussalam"                                   
#> [369] "Colombia"                                            
#> [370] "Hungary"                                             
#> [371] "Egypt"                                               
#> [372] "Canada"                                              
#> [373] "Cayman Islands"                                      
#> [374] "Argentina"                                           
#> [375] "Jamaica"                                             
#> [376] "Kuwait"                                              
#> [377] "Cambodia"                                            
#> [378] "Sao Tome and Principe"                               
#> [379] "Isle of Man"                                         
#> [380] "Saint Kitts and Nevis"                               
#> [381] "Zimbabwe"                                            
#> [382] "Liechtenstein"                                       
#> [383] "North Macedonia"                                     
#> [384] "Burundi"                                             
#> [385] "Cyprus"                                              
#> [386] "Grenada"                                             
#> [387] "Finland"                                             
#> [388] "Martinique"                                          
#> [389] "Belgium"                                             
#> [390] "Guadeloupe"                                          
#> [391] "Cayman Islands"                                      
#> [392] "Paraguay"                                            
#> [393] "Sudan"                                               
#> [394] "Turks and Caicos Islands"                            
#> [395] "Belize"                                              
#> [396] "Saint Martin (French part)"                          
#> [397] "Slovakia"                                            
#> [398] "Virgin Islands (U.S.)"                               
#> [399] "Mozambique"                                          
#> [400] "Wallis and Futuna"                                   
#> [401] "Bangladesh"                                          
#> [402] "Azerbaijan"                                          
#> [403] "Lebanon"                                             
#> [404] "Saint Martin (French part)"                          
#> [405] "United Kingdom of Great Britain and Northern Ireland"
#> [406] "Cameroon"                                            
#> [407] "Ghana"                                               
#> [408] "Venezuela (Bolivarian Republic of)"                  
#> [409] "Cabo Verde"                                          
#> [410] "Palau"                                               
#> [411] "Ecuador"                                             
#> [412] "Saint Pierre and Miquelon"                           
#> [413] "Falkland Islands (Malvinas)"                         
#> [414] "Viet Nam"                                            
#> [415] "Belgium"                                             
#> [416] "French Guiana"                                       
#> [417] "Cook Islands"                                        
#> [418] "Cyprus"                                              
#> [419] "Pitcairn"                                            
#> [420] "Gibraltar"                                           
#> [421] "Lao People's Democratic Republic"                    
#> [422] "Heard Island and McDonald Islands"                   
#> [423] "Belize"                                              
#> [424] "Bahamas"                                             
#> [425] "Northern Mariana Islands"                            
#> [426] "Falkland Islands (Malvinas)"                         
#> [427] "Andorra"                                             
#> [428] "Liechtenstein"                                       
#> [429] "Ethiopia"                                            
#> [430] "United Arab Emirates"                                
#> [431] "Cayman Islands"                                      
#> [432] "Nauru"                                               
#> [433] "Mongolia"                                            
#> [434] "New Zealand"                                         
#> [435] "Poland"                                              
#> [436] "Nepal"                                               
#> [437] "Peru"                                                
#> [438] "Comoros"                                             
#> [439] "Spain"                                               
#> [440] "Guinea-Bissau"                                       
#> [441] "Guadeloupe"                                          
#> [442] "Afghanistan"                                         
#> [443] "Central African Republic"                            
#> [444] "Ukraine"                                             
#> [445] "British Indian Ocean Territory"                      
#> [446] "Liechtenstein"                                       
#> [447] "Benin"                                               
#> [448] "Namibia"                                             
#> [449] "Seychelles"                                          
#> [450] "Mauritania"                                          
#> [451] "Lebanon"                                             
#> [452] "Myanmar"                                             
#> [453] "Philippines"                                         
#> [454] "Cabo Verde"                                          
#> [455] "Luxembourg"                                          
#> [456] "Netherlands"                                         
#> [457] "Latvia"                                              
#> [458] "Senegal"                                             
#> [459] "Azerbaijan"                                          
#> [460] "Mauritania"                                          
#> [461] "China"                                               
#> [462] "Thailand"                                            
#> [463] "Morocco"                                             
#> [464] "Comoros"                                             
#> [465] "Uruguay"                                             
#> [466] "Sint Maarten (Dutch part)"                           
#> [467] "Papua New Guinea"                                    
#> [468] "Argentina"                                           
#> [469] "Mayotte"                                             
#> [470] "Congo, Democratic Republic of the"                   
#> [471] "Guinea-Bissau"                                       
#> [472] "French Southern Territories"                         
#> [473] "Liberia"                                             
#> [474] "Micronesia (Federated States of)"                    
#> [475] "Guam"                                                
#> [476] "Norfolk Island"                                      
#> [477] "Taiwan, Province of China"                           
#> [478] "Russian Federation"                                  
#> [479] "Slovakia"                                            
#> [480] "Latvia"                                              
#> [481] "Nauru"                                               
#> [482] "Brunei Darussalam"                                   
#> [483] "Nigeria"                                             
#> [484] "Colombia"                                            
#> [485] "Åland Islands"                                       
#> [486] "Cyprus"                                              
#> [487] "Guam"                                                
#> [488] "Bonaire, Sint Eustatius and Saba"                    
#> [489] "Saint Pierre and Miquelon"                           
#> [490] "Montserrat"                                          
#> [491] "Moldova, Republic of"                                
#> [492] "Chad"                                                
#> [493] "Marshall Islands"                                    
#> [494] "Romania"                                             
#> [495] "Panama"                                              
#> [496] "Finland"                                             
#> [497] "South Sudan"                                         
#> [498] "Montserrat"                                          
#> [499] "Lao People's Democratic Republic"                    
#> [500] "Ecuador"                                             
random_countries(n = 5, nomenclature = "ISO3", seed = 5)
#> [1] "DJI" "IND" "RWA" "ZAF" "SVK"