Output random country names

This function returns the mode of vectors. That is to say, for any given vector of values, it returns the value that appears most frequently. The function works with strings, numerical and mixed inputs. NA values are treated as distinct values.

Usage

random_countries(n, replace = FALSE, nomenclature = "name_en", seed = NULL)

Arguments

n

Number of desired (pseudo)random country names.

replace

Logical value indicating whether sampling should be with replacement.

nomenclature

Nomenclature from which the list of countries should be taken. Not all countries are present in all nomenclature, for example Taiwan is not recognised by the UN, so it will not be returned with "WTO_en". The function accept any of the nomenclatures of country_name. For a list of accepted values, refer to this page. The default is name_en, which is the English list of names in the ISO standard 3166.

seed

Single numerical value to be used as seed.

Value

A vector of n (pseudo)random country names.

See also

Examples

random_countries(10)
#>  [1] "Myanmar"                  "El Salvador"             
#>  [3] "Virgin Islands (British)" "Malta"                   
#>  [5] "Svalbard and Jan Mayen"   "Indonesia"               
#>  [7] "Botswana"                 "Tunisia"                 
#>  [9] "Israel"                   "Lithuania"               
random_countries(n = 500, replace = TRUE)
#>   [1] "Maldives"                                            
#>   [2] "Slovakia"                                            
#>   [3] "Cabo Verde"                                          
#>   [4] "Belarus"                                             
#>   [5] "Costa Rica"                                          
#>   [6] "Singapore"                                           
#>   [7] "Grenada"                                             
#>   [8] "Virgin Islands (British)"                            
#>   [9] "Turkey"                                              
#>  [10] "Suriname"                                            
#>  [11] "Iceland"                                             
#>  [12] "Malta"                                               
#>  [13] "Ethiopia"                                            
#>  [14] "Canada"                                              
#>  [15] "Bahamas"                                             
#>  [16] "Grenada"                                             
#>  [17] "Croatia"                                             
#>  [18] "Latvia"                                              
#>  [19] "Indonesia"                                           
#>  [20] "Afghanistan"                                         
#>  [21] "Bosnia and Herzegovina"                              
#>  [22] "Guadeloupe"                                          
#>  [23] "Vanuatu"                                             
#>  [24] "Andorra"                                             
#>  [25] "North Macedonia"                                     
#>  [26] "Greece"                                              
#>  [27] "Niue"                                                
#>  [28] "Slovakia"                                            
#>  [29] "Viet Nam"                                            
#>  [30] "Cambodia"                                            
#>  [31] "Congo"                                               
#>  [32] "Grenada"                                             
#>  [33] "Equatorial Guinea"                                   
#>  [34] "South Georgia and the South Sandwich Islands"        
#>  [35] "Croatia"                                             
#>  [36] "Belarus"                                             
#>  [37] "Isle of Man"                                         
#>  [38] "China"                                               
#>  [39] "Slovakia"                                            
#>  [40] "Virgin Islands (British)"                            
#>  [41] "Wallis and Futuna"                                   
#>  [42] "Thailand"                                            
#>  [43] "Samoa"                                               
#>  [44] "Ireland"                                             
#>  [45] "Timor-Leste"                                         
#>  [46] "Christmas Island"                                    
#>  [47] "Guinea-Bissau"                                       
#>  [48] "Cocos (Keeling) Islands"                             
#>  [49] "Botswana"                                            
#>  [50] "Slovakia"                                            
#>  [51] "Equatorial Guinea"                                   
#>  [52] "Martinique"                                          
#>  [53] "Solomon Islands"                                     
#>  [54] "British Indian Ocean Territory"                      
#>  [55] "Israel"                                              
#>  [56] "Bahrain"                                             
#>  [57] "Cocos (Keeling) Islands"                             
#>  [58] "Saint Barthélemy"                                    
#>  [59] "Turkmenistan"                                        
#>  [60] "Dominican Republic"                                  
#>  [61] "Vanuatu"                                             
#>  [62] "Guam"                                                
#>  [63] "Palau"                                               
#>  [64] "Oman"                                                
#>  [65] "Samoa"                                               
#>  [66] "Turkey"                                              
#>  [67] "Saint Barthélemy"                                    
#>  [68] "Bangladesh"                                          
#>  [69] "Montenegro"                                          
#>  [70] "Kyrgyzstan"                                          
#>  [71] "Greece"                                              
#>  [72] "Guatemala"                                           
#>  [73] "Aruba"                                               
#>  [74] "Zimbabwe"                                            
#>  [75] "Congo, Democratic Republic of the"                   
#>  [76] "Czechia"                                             
#>  [77] "Nauru"                                               
#>  [78] "Netherlands"                                         
#>  [79] "Guam"                                                
#>  [80] "Cyprus"                                              
#>  [81] "North Macedonia"                                     
#>  [82] "Albania"                                             
#>  [83] "Bangladesh"                                          
#>  [84] "Rwanda"                                              
#>  [85] "Barbados"                                            
#>  [86] "New Zealand"                                         
#>  [87] "Canada"                                              
#>  [88] "Tokelau"                                             
#>  [89] "Guadeloupe"                                          
#>  [90] "Jersey"                                              
#>  [91] "Togo"                                                
#>  [92] "Palestine, State of"                                 
#>  [93] "Nepal"                                               
#>  [94] "Falkland Islands (Malvinas)"                         
#>  [95] "Lao People's Democratic Republic"                    
#>  [96] "Austria"                                             
#>  [97] "Norfolk Island"                                      
#>  [98] "Slovenia"                                            
#>  [99] "Mauritania"                                          
#> [100] "Saint Lucia"                                         
#> [101] "Zimbabwe"                                            
#> [102] "Norway"                                              
#> [103] "Guinea-Bissau"                                       
#> [104] "Georgia"                                             
#> [105] "Nigeria"                                             
#> [106] "Luxembourg"                                          
#> [107] "Peru"                                                
#> [108] "American Samoa"                                      
#> [109] "South Africa"                                        
#> [110] "Jamaica"                                             
#> [111] "Jamaica"                                             
#> [112] "Monaco"                                              
#> [113] "Haiti"                                               
#> [114] "Mongolia"                                            
#> [115] "Cambodia"                                            
#> [116] "Finland"                                             
#> [117] "Venezuela (Bolivarian Republic of)"                  
#> [118] "Algeria"                                             
#> [119] "Marshall Islands"                                    
#> [120] "Belize"                                              
#> [121] "Costa Rica"                                          
#> [122] "Equatorial Guinea"                                   
#> [123] "Jordan"                                              
#> [124] "Western Sahara"                                      
#> [125] "French Polynesia"                                    
#> [126] "Cambodia"                                            
#> [127] "Vanuatu"                                             
#> [128] "Niue"                                                
#> [129] "Spain"                                               
#> [130] "Saint Lucia"                                         
#> [131] "Saint Pierre and Miquelon"                           
#> [132] "Kenya"                                               
#> [133] "Kenya"                                               
#> [134] "Isle of Man"                                         
#> [135] "Wallis and Futuna"                                   
#> [136] "Italy"                                               
#> [137] "Viet Nam"                                            
#> [138] "Svalbard and Jan Mayen"                              
#> [139] "Peru"                                                
#> [140] "Morocco"                                             
#> [141] "Palestine, State of"                                 
#> [142] "Saint Lucia"                                         
#> [143] "French Southern Territories"                         
#> [144] "Bhutan"                                              
#> [145] "Russian Federation"                                  
#> [146] "Hong Kong"                                           
#> [147] "Martinique"                                          
#> [148] "Zambia"                                              
#> [149] "Indonesia"                                           
#> [150] "United States Minor Outlying Islands"                
#> [151] "American Samoa"                                      
#> [152] "Madagascar"                                          
#> [153] "Central African Republic"                            
#> [154] "Austria"                                             
#> [155] "Virgin Islands (British)"                            
#> [156] "Poland"                                              
#> [157] "Saint Barthélemy"                                    
#> [158] "Syrian Arab Republic"                                
#> [159] "Djibouti"                                            
#> [160] "Timor-Leste"                                         
#> [161] "Anguilla"                                            
#> [162] "Liberia"                                             
#> [163] "Guinea-Bissau"                                       
#> [164] "Latvia"                                              
#> [165] "Serbia"                                              
#> [166] "Aruba"                                               
#> [167] "Eswatini"                                            
#> [168] "Marshall Islands"                                    
#> [169] "Singapore"                                           
#> [170] "San Marino"                                          
#> [171] "Nicaragua"                                           
#> [172] "South Sudan"                                         
#> [173] "Georgia"                                             
#> [174] "Philippines"                                         
#> [175] "Ukraine"                                             
#> [176] "Lebanon"                                             
#> [177] "Antigua and Barbuda"                                 
#> [178] "Japan"                                               
#> [179] "Mozambique"                                          
#> [180] "Barbados"                                            
#> [181] "Bouvet Island"                                       
#> [182] "Korea (Democratic People's Republic of)"             
#> [183] "Tunisia"                                             
#> [184] "Wallis and Futuna"                                   
#> [185] "Burkina Faso"                                        
#> [186] "Cambodia"                                            
#> [187] "Honduras"                                            
#> [188] "Iraq"                                                
#> [189] "Svalbard and Jan Mayen"                              
#> [190] "Ghana"                                               
#> [191] "Lao People's Democratic Republic"                    
#> [192] "Panama"                                              
#> [193] "Ukraine"                                             
#> [194] "Syrian Arab Republic"                                
#> [195] "Switzerland"                                         
#> [196] "Tanzania, United Republic of"                        
#> [197] "Serbia"                                              
#> [198] "Central African Republic"                            
#> [199] "Guernsey"                                            
#> [200] "Uzbekistan"                                          
#> [201] "Georgia"                                             
#> [202] "Algeria"                                             
#> [203] "Brazil"                                              
#> [204] "Paraguay"                                            
#> [205] "Czechia"                                             
#> [206] "United States Minor Outlying Islands"                
#> [207] "Tajikistan"                                          
#> [208] "Marshall Islands"                                    
#> [209] "Côte d'Ivoire"                                       
#> [210] "Belarus"                                             
#> [211] "Philippines"                                         
#> [212] "Cambodia"                                            
#> [213] "Israel"                                              
#> [214] "Myanmar"                                             
#> [215] "Peru"                                                
#> [216] "Benin"                                               
#> [217] "United States of America"                            
#> [218] "Uruguay"                                             
#> [219] "Slovakia"                                            
#> [220] "Norfolk Island"                                      
#> [221] "Bosnia and Herzegovina"                              
#> [222] "Pakistan"                                            
#> [223] "Virgin Islands (British)"                            
#> [224] "Jersey"                                              
#> [225] "India"                                               
#> [226] "Oman"                                                
#> [227] "Norway"                                              
#> [228] "Faroe Islands"                                       
#> [229] "Costa Rica"                                          
#> [230] "Angola"                                              
#> [231] "Iraq"                                                
#> [232] "Serbia"                                              
#> [233] "Virgin Islands (British)"                            
#> [234] "Bouvet Island"                                       
#> [235] "Haiti"                                               
#> [236] "Saint Helena, Ascension and Tristan da Cunha"        
#> [237] "Turkmenistan"                                        
#> [238] "Canada"                                              
#> [239] "Iceland"                                             
#> [240] "Canada"                                              
#> [241] "Turkey"                                              
#> [242] "Bahamas"                                             
#> [243] "Armenia"                                             
#> [244] "Jordan"                                              
#> [245] "Philippines"                                         
#> [246] "Saint Pierre and Miquelon"                           
#> [247] "Tuvalu"                                              
#> [248] "Saint Helena, Ascension and Tristan da Cunha"        
#> [249] "Cambodia"                                            
#> [250] "Antigua and Barbuda"                                 
#> [251] "Guatemala"                                           
#> [252] "Belarus"                                             
#> [253] "Uzbekistan"                                          
#> [254] "Saint Lucia"                                         
#> [255] "Senegal"                                             
#> [256] "Turks and Caicos Islands"                            
#> [257] "Ireland"                                             
#> [258] "United States of America"                            
#> [259] "Qatar"                                               
#> [260] "French Southern Territories"                         
#> [261] "Comoros"                                             
#> [262] "Guadeloupe"                                          
#> [263] "Zambia"                                              
#> [264] "Lesotho"                                             
#> [265] "Tunisia"                                             
#> [266] "Georgia"                                             
#> [267] "United Arab Emirates"                                
#> [268] "Austria"                                             
#> [269] "Virgin Islands (British)"                            
#> [270] "Greece"                                              
#> [271] "Egypt"                                               
#> [272] "Zambia"                                              
#> [273] "Mayotte"                                             
#> [274] "Lesotho"                                             
#> [275] "Jamaica"                                             
#> [276] "Romania"                                             
#> [277] "Aruba"                                               
#> [278] "India"                                               
#> [279] "Mongolia"                                            
#> [280] "Oman"                                                
#> [281] "Gabon"                                               
#> [282] "Isle of Man"                                         
#> [283] "Cayman Islands"                                      
#> [284] "French Southern Territories"                         
#> [285] "Kenya"                                               
#> [286] "Poland"                                              
#> [287] "Togo"                                                
#> [288] "Nauru"                                               
#> [289] "United Kingdom of Great Britain and Northern Ireland"
#> [290] "Liechtenstein"                                       
#> [291] "San Marino"                                          
#> [292] "Greenland"                                           
#> [293] "Belgium"                                             
#> [294] "Congo"                                               
#> [295] "Syrian Arab Republic"                                
#> [296] "Armenia"                                             
#> [297] "Dominican Republic"                                  
#> [298] "Armenia"                                             
#> [299] "Netherlands"                                         
#> [300] "Italy"                                               
#> [301] "Portugal"                                            
#> [302] "United Kingdom of Great Britain and Northern Ireland"
#> [303] "Bhutan"                                              
#> [304] "Liberia"                                             
#> [305] "Algeria"                                             
#> [306] "Tokelau"                                             
#> [307] "Ghana"                                               
#> [308] "South Sudan"                                         
#> [309] "Kazakhstan"                                          
#> [310] "Algeria"                                             
#> [311] "Angola"                                              
#> [312] "Mali"                                                
#> [313] "Lesotho"                                             
#> [314] "Ethiopia"                                            
#> [315] "Argentina"                                           
#> [316] "Brazil"                                              
#> [317] "Ghana"                                               
#> [318] "Congo"                                               
#> [319] "Canada"                                              
#> [320] "Cameroon"                                            
#> [321] "South Africa"                                        
#> [322] "Vanuatu"                                             
#> [323] "Kuwait"                                              
#> [324] "Antarctica"                                          
#> [325] "Malta"                                               
#> [326] "United States of America"                            
#> [327] "Equatorial Guinea"                                   
#> [328] "Isle of Man"                                         
#> [329] "Chad"                                                
#> [330] "Isle of Man"                                         
#> [331] "Zambia"                                              
#> [332] "Bulgaria"                                            
#> [333] "Dominican Republic"                                  
#> [334] "French Southern Territories"                         
#> [335] "Congo"                                               
#> [336] "Kazakhstan"                                          
#> [337] "Tanzania, United Republic of"                        
#> [338] "French Southern Territories"                         
#> [339] "Estonia"                                             
#> [340] "Rwanda"                                              
#> [341] "Austria"                                             
#> [342] "Kiribati"                                            
#> [343] "Afghanistan"                                         
#> [344] "Australia"                                           
#> [345] "Paraguay"                                            
#> [346] "Burkina Faso"                                        
#> [347] "Sao Tome and Principe"                               
#> [348] "El Salvador"                                         
#> [349] "Slovakia"                                            
#> [350] "Samoa"                                               
#> [351] "Nauru"                                               
#> [352] "Greenland"                                           
#> [353] "Latvia"                                              
#> [354] "Palestine, State of"                                 
#> [355] "Montenegro"                                          
#> [356] "Venezuela (Bolivarian Republic of)"                  
#> [357] "Samoa"                                               
#> [358] "Andorra"                                             
#> [359] "Guinea"                                              
#> [360] "Singapore"                                           
#> [361] "Bolivia (Plurinational State of)"                    
#> [362] "Ireland"                                             
#> [363] "Benin"                                               
#> [364] "Faroe Islands"                                       
#> [365] "Gibraltar"                                           
#> [366] "Cayman Islands"                                      
#> [367] "Guinea"                                              
#> [368] "Marshall Islands"                                    
#> [369] "Tanzania, United Republic of"                        
#> [370] "Seychelles"                                          
#> [371] "Guadeloupe"                                          
#> [372] "Brunei Darussalam"                                   
#> [373] "Israel"                                              
#> [374] "Moldova, Republic of"                                
#> [375] "Western Sahara"                                      
#> [376] "Malta"                                               
#> [377] "Somalia"                                             
#> [378] "Syrian Arab Republic"                                
#> [379] "Mexico"                                              
#> [380] "Romania"                                             
#> [381] "Cayman Islands"                                      
#> [382] "Macao"                                               
#> [383] "Côte d'Ivoire"                                       
#> [384] "Turkmenistan"                                        
#> [385] "Guatemala"                                           
#> [386] "United Arab Emirates"                                
#> [387] "Western Sahara"                                      
#> [388] "Costa Rica"                                          
#> [389] "Holy See"                                            
#> [390] "Venezuela (Bolivarian Republic of)"                  
#> [391] "Samoa"                                               
#> [392] "Tonga"                                               
#> [393] "Norway"                                              
#> [394] "Iraq"                                                
#> [395] "Cameroon"                                            
#> [396] "Tanzania, United Republic of"                        
#> [397] "Argentina"                                           
#> [398] "Saint Vincent and the Grenadines"                    
#> [399] "Saint Pierre and Miquelon"                           
#> [400] "Djibouti"                                            
#> [401] "Sierra Leone"                                        
#> [402] "Virgin Islands (British)"                            
#> [403] "Tuvalu"                                              
#> [404] "Guatemala"                                           
#> [405] "Madagascar"                                          
#> [406] "Mexico"                                              
#> [407] "Iran (Islamic Republic of)"                          
#> [408] "Poland"                                              
#> [409] "Djibouti"                                            
#> [410] "Azerbaijan"                                          
#> [411] "Comoros"                                             
#> [412] "Dominica"                                            
#> [413] "South Africa"                                        
#> [414] "Sweden"                                              
#> [415] "China"                                               
#> [416] "French Guiana"                                       
#> [417] "Venezuela (Bolivarian Republic of)"                  
#> [418] "Curaçao"                                             
#> [419] "Mali"                                                
#> [420] "Guinea"                                              
#> [421] "Djibouti"                                            
#> [422] "Comoros"                                             
#> [423] "Eswatini"                                            
#> [424] "New Caledonia"                                       
#> [425] "Palestine, State of"                                 
#> [426] "Solomon Islands"                                     
#> [427] "Guernsey"                                            
#> [428] "Hong Kong"                                           
#> [429] "French Polynesia"                                    
#> [430] "Chile"                                               
#> [431] "Poland"                                              
#> [432] "Chile"                                               
#> [433] "Panama"                                              
#> [434] "Cook Islands"                                        
#> [435] "Saint Lucia"                                         
#> [436] "Sao Tome and Principe"                               
#> [437] "Somalia"                                             
#> [438] "Guam"                                                
#> [439] "French Southern Territories"                         
#> [440] "Iran (Islamic Republic of)"                          
#> [441] "Serbia"                                              
#> [442] "Curaçao"                                             
#> [443] "Zimbabwe"                                            
#> [444] "Niue"                                                
#> [445] "Cyprus"                                              
#> [446] "France"                                              
#> [447] "Nauru"                                               
#> [448] "Tajikistan"                                          
#> [449] "Belgium"                                             
#> [450] "Taiwan, Province of China"                           
#> [451] "Nigeria"                                             
#> [452] "Bolivia (Plurinational State of)"                    
#> [453] "Brunei Darussalam"                                   
#> [454] "Saint Vincent and the Grenadines"                    
#> [455] "Mauritius"                                           
#> [456] "Viet Nam"                                            
#> [457] "New Zealand"                                         
#> [458] "Singapore"                                           
#> [459] "Eswatini"                                            
#> [460] "Gibraltar"                                           
#> [461] "Wallis and Futuna"                                   
#> [462] "Palestine, State of"                                 
#> [463] "Monaco"                                              
#> [464] "Moldova, Republic of"                                
#> [465] "Taiwan, Province of China"                           
#> [466] "Portugal"                                            
#> [467] "Israel"                                              
#> [468] "Libya"                                               
#> [469] "Christmas Island"                                    
#> [470] "Kuwait"                                              
#> [471] "Mexico"                                              
#> [472] "Pitcairn"                                            
#> [473] "Heard Island and McDonald Islands"                   
#> [474] "Venezuela (Bolivarian Republic of)"                  
#> [475] "Cook Islands"                                        
#> [476] "Belgium"                                             
#> [477] "Faroe Islands"                                       
#> [478] "Cuba"                                                
#> [479] "Equatorial Guinea"                                   
#> [480] "Kenya"                                               
#> [481] "Israel"                                              
#> [482] "Monaco"                                              
#> [483] "Mali"                                                
#> [484] "Jersey"                                              
#> [485] "South Sudan"                                         
#> [486] "Malta"                                               
#> [487] "Palestine, State of"                                 
#> [488] "South Africa"                                        
#> [489] "Bahrain"                                             
#> [490] "Slovenia"                                            
#> [491] "Haiti"                                               
#> [492] "Lao People's Democratic Republic"                    
#> [493] "Brazil"                                              
#> [494] "Bhutan"                                              
#> [495] "Faroe Islands"                                       
#> [496] "Malawi"                                              
#> [497] "Croatia"                                             
#> [498] "Italy"                                               
#> [499] "Guatemala"                                           
#> [500] "New Caledonia"                                       
random_countries(n = 5, nomenclature = "ISO3", seed = 5)
#> [1] "DJI" "IND" "RWA" "ZAF" "SVK"